### 2022-12-08 #4998 @editorialbot
[PRE REVIEW]: CORA: A Python Package for Combinational Regularity Analysis
**Submitting author:** @ZuzanaSebb (<a href="http://orcid.org/0000-0002-4669-5659">Zuzana Sebechlebská</a>)
**Repository:** https://github.com/PoliUniLu/cora
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.1
