### 2022-12-08 #4998 @danielskatz
👋 @ZuzanaSebb - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and note that you can click on the error above to find out more about it. Please use the command `@editorialbot generate pdf` after making changes to the .md file to make a new PDF. editorialbot commands need to be the first entry in a new comment.  Once the paper builds, you should also use the command `@editorialbot check references` to check the references.
