### 2022-12-11 cb1fb5c
Update paper.md
Citation fix
