+++ paper/paper.md
@@ -121,7 +121,7 @@ Logic diagrams are graphical representations of Boolean-algebraic
 functions that have so far almost exclusively been used in electrical
 engineering. Over the last 10 years, however, scientists from other
 disciplines have begun to discover the utility of logic diagrams
-`\citep[e.g.,][]{delorenzo_schmidt_2018, pearl_2009, thiem_et_al_2020}`{=latex}.
+[@delorenzo_schmidt_2018; @pearl_2009; @thiem_et_al_2020].
 The `LOGIGRAM` package in `CORA` is an interactive tool for the
 standardized production of two-level logic diagrams, called "logigrams"
 in CORA. Two examples of a logigram, including the interface through
