### 2022-12-11 7d718eb
Update paper.bib
Add missing DOI-s
