+++ paper/paper.bib
@@ -11,6 +11,7 @@
    journal = {Sociological Methods \& Research},
    volume = {38},
    number = {1},
+   doi = {10.1177/0049124109339369},
    pages = {71-101}
 }
 @book{cronqvist_2019,
@@ -26,7 +27,8 @@
    title = {Biological standards for the Knowledge-Based BioEconomy: What is at stake},
    journal = {New Biotechnology},
    volume = {40},
-   pages = {170-180}
+   pages = {170-180},
+   doi = {10.1016/j.nbt.2017.05.001}
 }
 @book{dusa_2021,
    author = {Dusa, Adrian},
@@ -57,7 +59,7 @@
    title = {Causality: Models, Reasoning, and Inference},
    publisher = {Cambridge University Press},
    address = {Cambridge},
-   edition = {2nd}
+   edition = {2nd},
 }
 @book{ragin_1987,
    author = {Ragin, Charles C.},
@@ -79,7 +81,8 @@
    journal = {Health Psychology},
    volume = {38},
    number = {9},
-   pages = {769-771}
+   pages = {769-771},
+   doi = {10.1037/hea0000783},
 }
 @book{thiem_2018,
    author = {Thiem, Alrik},
@@ -93,7 +96,8 @@
    Volume = {15},
    Number = {6},
    Title = {{Algorithmic bias in social research: A meta-analysis}},
-   Year = 2020}
+   Year = 2020},
+   doi = {10.1371/journal.pone.0233625},
    
 @article{thiem_et_al_forth,
    author = {Thiem, Alrik and Mkrtchyan, Lusine and Sebechlebská, Zuzana},
