### 2022-12-11 d865fbb
Update paper.bib
Fix syntax error in .bib file
