+++ paper/paper.bib
@@ -96,7 +96,7 @@
    Volume = {15},
    Number = {6},
    Title = {{Algorithmic bias in social research: A meta-analysis}},
-   Year = 2020},
+   Year = {2020},
    doi = {10.1371/journal.pone.0233625},
    
 @article{thiem_et_al_forth,
