+++ paper/paper.bib
@@ -101,9 +101,10 @@
    
 @article{thiem_et_al_forth,
    author = {Thiem, Alrik and Mkrtchyan, Lusine and Sebechlebská, Zuzana},
-   year = {in press},
-   title = {{Combinational Regularity Analysis (CORA)}: A New Method for Uncovering Complex Causation in Medical and Health Research},
    journal = {BMC Medical Research Methodology},
+   year = {2022},
+   title = {{Combinational Regularity Analysis (CORA)}: A New Method for Uncovering Complex Causation in Medical and Health Research},
+   
 }
 
 @article{thiem_dusa_2013,
