+++ paper/paper.bib
@@ -98,7 +98,7 @@
    Title = {{Algorithmic bias in social research: A meta-analysis}},
    Year = {2020},
    doi = {10.1371/journal.pone.0233625},
-   
+}  
 @article{thiem_et_al_forth,
    author = {Thiem, Alrik and Mkrtchyan, Lusine and Sebechlebská, Zuzana},
    journal = {BMC Medical Research Methodology},
