### 2022-12-12 42c698a
Correction in the figures.
