### 2022-12-12 a1d151b
Fix in the figures.
