+++ paper/paper.bib
@@ -11,8 +11,8 @@
    journal = {Sociological Methods \& Research},
    volume = {38},
    number = {1},
-   doi = {10.1177/0049124109339369},
-   pages = {71-101}
+   pages = {71-101},
+   doi = {10.1177/0049124109339369}
 }
 @book{cronqvist_2019,
    author = {Cronqvist, Lasse},
@@ -91,40 +91,28 @@
    publisher = {http://cran.r-project.org/package=QCApro},
 }
 @article{thiem_et_al_2020,
-   Author = {{Thiem}, A. and {Mkrtchyan}, L. and {Haesebrouck}, T. and {Sanchez}, D.},
-   Journal = {PLoS ONE},
-   Volume = {15},
-   Number = {6},
-   Title = {{Algorithmic bias in social research: A meta-analysis}},
-   Year = {2020},
-   doi = {10.1371/journal.pone.0233625},
-}  
+   author = {Thiem, Alrik and Mkrtchyan, Lusine and Haesebrouck, Tim and Sanchez, David},
+   year = {2020},
+   title = {Algorithmic bias in social research: A meta-analysis},
+   journal = {PLOS ONE},
+   volume = {15},
+   number = {6},
+   pages = {e0233625},
+   doi = {10.1371/journal.pone.0233625}
+}
 @article{thiem_et_al_forth,
    author = {Thiem, Alrik and Mkrtchyan, Lusine and Sebechlebská, Zuzana},
+   year = {forthcoming},
    journal = {BMC Medical Research Methodology},
-   year = {2022},
-   title = {{Combinational Regularity Analysis (CORA)}: A New Method for Uncovering Complex Causation in Medical and Health Research},
-   
+   title = {{Combinational Regularity Analysis (CORA)} - A new method for uncovering complex causation in medical and health research},  
 }
-
 @article{thiem_dusa_2013,
    author = {Thiem, Alrik and Du\c{s}a, Adrian},
    year = {2013},
-   title = {Boolean Minimization in Social Science Research: A Review of Current Software for {Qualitative Comparative Analysis (QCA)}},
+   title = {Boolean minimization in social science research: A review of current software for {Qualitative Comparative Analysis (QCA)}},
    journal = {Social Science Computer Review},
    volume = {31},
    number = {4},
-   pages = {505-521}
+   pages = {505-521},
+   doi = {10.1177/0894439313478999}
 }
-
-@article{Pearson:2017,
-	Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
-	Adsurl = {http://adsabs.harvard.edu/abs/2017arXiv170304627P},
-	Archiveprefix = {arXiv},
-	Author = {{Pearson}, S. and {Price-Whelan}, A.~M. and {Johnston}, K.~V.},
-	Eprint = {1703.04627},
-	Journal = {ArXiv e-prints},
-	Keywords = {Astrophysics - Astrophysics of Galaxies},
-	Month = mar,
-	Title = {{Gaps in Globular Cluster Streams: Pal 5 and the Galactic Bar}},
-	Year = 2017}
