+++ paper/paper.bib
@@ -1,11 +1,11 @@
 @book{ambuhl_baumgartner_2021,
-   author = {Ambuehl, Mathias and Baumgartner, Michael},
+   author = {Ambuehl, M. and Baumgartner, M.},
    year = {2020},
    title = {cna: Causal Modeling with Coincidence Analysis, Version 3.4.0},
    publisher = {http://cran.r-project.org/package=cna}
 }
 @article{baumgartner_2009,
-   author = {Baumgartner, Michael},
+   author = {Baumgartner, M.},
    year = {2009},
    title = {Inferring Causal Complexity},
    journal = {Sociological Methods \& Research},
@@ -15,14 +15,14 @@
    doi = {10.1177/0049124109339369}
 }
 @book{cronqvist_2019,
-   author = {Cronqvist, Lasse},
+   author = {Cronqvist, L.},
    year = {2019},
    title = {Tosmana: Tool for Small-N Analysis, Version 1.61 [Computer Program]},
    publisher = {University of Trier},
    address = {Trier},
 }
 @article{delorenzo_schmidt_2018,
-   author = {de Lorenzo, Víctor and Schmidt, Markus},
+   author = {de Lorenzo, V. and Schmidt, M.},
    year = {2018},
    title = {Biological standards for the Knowledge-Based BioEconomy: What is at stake},
    journal = {New Biotechnology},
@@ -31,13 +31,13 @@
    doi = {10.1016/j.nbt.2017.05.001}
 }
 @book{dusa_2021,
-   author = {Dusa, Adrian},
+   author = {Dusa, A.},
    year = {2021},
    title = {QCA: Qualitative Comparative Analysis, Version 3.11},
    publisher = {https://cran.r-project.org/web/packages/QCA/},
 }
 @book{lewin_protheroe_1992,
-   author = {Lewin, Douglas and Protheroe, David},
+   author = {Lewin, D. and Protheroe, D.},
    year = {1992},
    title = {Design of Logic Systems},
    publisher = {Chapman \& Hall},
@@ -45,7 +45,7 @@
    edition = {2nd}
 }
 @article{mackie_1965,
-   author = {Mackie, John L.},
+   author = {Mackie, J.L.},
    year = {1965},
    title = {Causes and Conditions},
    journal = {American Philosophical Quarterly},
@@ -54,7 +54,7 @@
    pages = {245-264}
 }
 @book{pearl_2009,
-   author = {Pearl, Judea},
+   author = {Pearl, J.},
    year = {2009},
    title = {Causality: Models, Reasoning, and Inference},
    publisher = {Cambridge University Press},
@@ -62,20 +62,20 @@
    edition = {2nd},
 }
 @book{ragin_1987,
-   author = {Ragin, Charles C.},
+   author = {Ragin, C.C.},
    year = {1987},
    title = {The Comparative Method: Moving beyond Qualitative and Quantitative Strategies},
    publisher = {University of California Press},
    address = {Berkeley},
 }
 @book{ragin_davey_2019,
-   author = {Ragin, Charles C. and Davey, Sean},
+   author = {Ragin, C.C. and Davey, S.},
    year = {2019},
    title = {fs/QCA: Fuzzy-Set/Qualitative Comparative Analysis, Version 3.1b [Computer Program]},
    publisher = {http://www.socsci.uci.edu/~cragin/fsQCA/software.shtml},
 }
 @article{suls_green_2019,
-   author = {Suls, Jerry and Green, Paige A.},
+   author = {Suls, J. and Green, P.A.},
    year = {2019},
    title = {Multimorbidity in health psychology and behavioral medicine},
    journal = {Health Psychology},
@@ -85,13 +85,13 @@
    doi = {10.1037/hea0000783},
 }
[…]
