+++ paper/paper.md
@@ -7,7 +7,7 @@ tags:
   - combinational regularity analysis
   - switching circuit analysis
 authors:
-  - name: Zuzana Sebechlebska
+  - name: Zuzana Sebechlebská
     orcid: 0000-0002-4669-5659
     equal-contrib: true
     affiliation: "1" 
@@ -29,83 +29,24 @@ bibliography: paper.bib
 
 # Summary 
 
-Combinational Regularity Analysis (CORA) [@thiem_et_al_forth] is a new member
-of the family of configurational comparative methods (CCMs). Modern CCMs
-infer INUS structures from sets of configurational data. CORA
-generalizes the capabilities of existing CCMs through its ability to
-model multi-output data. It can thus infer INUS structures that include
-complex conjunctive effects. The accompanying and eponymous package
-`CORA` is an open-source Python tool, hosted on GitHub and deployed on
-the Google Colab platform.
+Combinational Regularity Analysis (CORA) [@thiem_et_al_forth] is a new member of the family of configurational comparative methods (CCMs). Modern CCMs infer INUS structures from sets of configurational data. CORA generalizes the capabilities of existing CCMs through its ability to model multi-output data. More precisely, CORA can infer INUS structures that include complex conjunctive effects. The accompanying and eponymous package `CORA` is an open-source Python tool, hosted on GitHub and deployed on the Google Colab platform.
 
 # Statement of need
 
-Modern configurational comparative methods (CCMs) can infer
-regularity-theoretic causal structures from categorical data. The most
-sophisticated of these structures are so-called INUS structures [@mackie_1965]. INUS structures are codified as two-level
-Boolean functions, such as $a\cdot b' + c' \Leftrightarrow z$, where
-$a$, $b$, $c$ and $z$ are Boolean literals, $\text{"} . \text{"}$ represents the
-Boolean multiplication operator, $\text{"} + \text{"}$ the Boolean addition operator,
-$\text{" } ' \text{ "}$ the Boolean negation operator, and $\text{"}\Leftrightarrow\text{"}$ the
-Boolean implication operator. Branches of Boolean algebra include, most
-importantly, set theory, propositional logic and switching circuit
-theory [@lewin_protheroe_1992].
+Modern CCMs can infer regularity-theoretic causal structures from categorical data. The most sophisticated of these structures are INUS structures, where "INUS" refers to an *i*nsufficient but *n*on-redundant part of a condition that is *u*nnecessary but *s*ufficient for some outcome [@mackie_1965]. INUS structures are syntactically codified as two-level Boolean functions, such as $a\cdot b' + c' \Leftrightarrow z$, where $a$, $b$, $c$ and $z$ are Boolean literals, $\text{"} \cdot \text{"}$ represents the Boolean multiplication operator, $\text{"} + \text{"}$ the Boolean addition operator, $\text{" } ' \text{ "}$ the Boolean negation operator, and $\text{"}\Leftrightarrow\text{"}$ the Boolean implication operator. Branches of Boolean algebra include, most importantly, set theory, propositional logic and switching circuit theory [@lewin_protheroe_1992].
 
-So far, the two most sophisticated CCMs have been Qualitative
-Comparative Analysis (QCA) [@ragin_1987] and
-Coincidence Analysis
-(CNA) [@baumgartner_2009]. For QCA, several
-software packages exist, each with different functionality
-[@cronqvist_2019, @dusa_2021, @ragin_davey_2019, @thiem_2018].
-However, since there is no harmonization on procedures in QCA, different
-software packages often generate different results
-[@thiem_dusa_2013]. For CNA, the `cna` package is
-available [@ambuhl_baumgartner_2021].
[…]
