+++ paper/paper.bib
@@ -1,30 +1,50 @@
 @book{ambuhl_baumgartner_2021,
    author = {Ambuehl, M. and Baumgartner, M.},
    year = {2020},
-   title = {cna: Causal Modeling with Coincidence Analysis, Version 3.4.0},
+   title = {{cna}: Causal modeling with {Coincidence Analysis}, version 3.4.0},
    publisher = {http://cran.r-project.org/package=cna}
 }
 @article{baumgartner_2009,
    author = {Baumgartner, M.},
    year = {2009},
-   title = {Inferring Causal Complexity},
+   title = {Inferring causal complexity},
    journal = {Sociological Methods \& Research},
    volume = {38},
    number = {1},
    pages = {71-101},
    doi = {10.1177/0049124109339369}
 }
+@article{baumgartner_thiem_2017,
+   author = {Baumgartner, M. and Thiem, A.},
+   year = {2017},
+   title = {Model Ambiguities in Configurational Comparative Research},
+   journal = {Sociological Methods \& Research},
+   volume = {46},
+   number = {4},
+   pages = {954-987},
+   doi = {10.1177/0049124115610351}
+}
+@article{baumgartner_thiem_2020,
+   author = {Baumgartner, M. and Thiem, A.},
+   year = {2020},
+   title = {Often Trusted but Never (Properly) Tested: Evaluating Qualitative Comparative Analysis},
+   journal = {Sociological Methods \& Research},
+   volume = {49},
+   number = {2},
+   pages = {279-311},
+   doi = {10.1177/0049124117701487}
+}
 @book{cronqvist_2019,
    author = {Cronqvist, L.},
    year = {2019},
-   title = {Tosmana: Tool for Small-N Analysis, Version 1.61 [Computer Program]},
+   title = {Tosmana: Tool for small-n analysis, version 1.61 [computer program]},
    publisher = {University of Trier},
    address = {Trier},
 }
 @article{delorenzo_schmidt_2018,
    author = {de Lorenzo, V. and Schmidt, M.},
    year = {2018},
-   title = {Biological standards for the Knowledge-Based BioEconomy: What is at stake},
+   title = {Biological standards for the knowledge-based bioeconomy: What is at stake},
    journal = {New Biotechnology},
    volume = {40},
    pages = {170-180},
@@ -33,13 +53,13 @@
 @book{dusa_2021,
    author = {Dusa, A.},
    year = {2021},
-   title = {QCA: Qualitative Comparative Analysis, Version 3.11},
+   title = {{QCA}: {Qualitative Comparative Analysis}, {R} package version 3.11},
    publisher = {https://cran.r-project.org/web/packages/QCA/},
 }
 @book{lewin_protheroe_1992,
    author = {Lewin, D. and Protheroe, D.},
    year = {1992},
-   title = {Design of Logic Systems},
+   title = {Design of logic systems},
    publisher = {Chapman \& Hall},
    address = {London},
    edition = {2nd}
@@ -47,7 +67,7 @@
 @article{mackie_1965,
    author = {Mackie, J.L.},
    year = {1965},
-   title = {Causes and Conditions},
+   title = {Causes and conditions},
    journal = {American Philosophical Quarterly},
    volume = {2},
    number = {4},
@@ -56,7 +76,7 @@
 @book{pearl_2009,
    author = {Pearl, J.},
    year = {2009},
-   title = {Causality: Models, Reasoning, and Inference},
+   title = {Causality: Models, reasoning, and inference},
    publisher = {Cambridge University Press},
    address = {Cambridge},
    edition = {2nd},
@@ -64,14 +84,14 @@
 @book{ragin_1987,
    author = {Ragin, C.C.},
    year = {1987},
[…]
