+++ paper/paper.md
@@ -29,15 +29,15 @@ bibliography: paper.bib
 
 # Summary 
 
-Combinational Regularity Analysis (CORA) [@thiem_et_al_forth] is a new member of the family of configurational comparative methods (CCMs). Modern CCMs infer INUS structures from sets of configurational data. CORA generalizes the capabilities of existing CCMs through its ability to model multi-output data. More precisely, CORA can infer INUS structures that include complex conjunctive effects. The accompanying and eponymous package `CORA` is an open-source Python tool, hosted on GitHub and deployed on the Google Colab platform.
+Combinational Regularity Analysis (CORA) [@thiem_et_al_forth] is a new member of the family of configurational comparative methods (CCMs). The most sophisticated CCMs infer INUS structures from sets of configurational data [@thiem_2022a]. CORA generalizes the capabilities of existing CCMs. By means of Boolean optimization algorithms for multi-output data, CORA can infer INUS structures that include complex conjunctive effects. The accompanying and eponymous package `CORA` is an open-source Python tool, hosted on GitHub and deployed on the Google Colab platform, with which such structures can be identified.
 
 # Statement of need
 
-Modern CCMs can infer regularity-theoretic causal structures from categorical data. The most sophisticated of these structures are INUS structures, where "INUS" refers to an *i*nsufficient but *n*on-redundant part of a condition that is *u*nnecessary but *s*ufficient for some outcome [@mackie_1965]. INUS structures are syntactically codified as two-level Boolean functions, such as $a\cdot b' + c' \Leftrightarrow z$, where $a$, $b$, $c$ and $z$ are Boolean literals, $\text{"} \cdot \text{"}$ represents the Boolean multiplication operator, $\text{"} + \text{"}$ the Boolean addition operator, $\text{" } ' \text{ "}$ the Boolean negation operator, and $\text{"}\Leftrightarrow\text{"}$ the Boolean implication operator. Branches of Boolean algebra include, most importantly, set theory, propositional logic and switching circuit theory [@lewin_protheroe_1992].
+Modern CCMs can infer regularity-theoretic causal structures from categorical data. The most sophisticated of these structures are so-called "INUS" structures. "INUS" refers to an *i*nsufficient but *n*on-redundant part of a condition that is *u*nnecessary but *s*ufficient for some outcome [@mackie_1965]. INUS structures are syntactically codified as two-level Boolean functions, such as $a\cdot b' + c' \Leftrightarrow z$, where $a$, $b$, $c$ and $z$ are Boolean literals, $\text{"} \cdot \text{"}$ represents the Boolean multiplication operator, $\text{"} + \text{"}$ the Boolean addition operator, $\text{" } ' \text{ "}$ the Boolean negation operator, and $\text{"}\Leftrightarrow\text{"}$ the Boolean implication operator. Branches of Boolean algebra include, most importantly, set theory, propositional logic and switching circuit theory [@lewin_protheroe_1992].
 
 So far, the two most sophisticated CCMs have been Qualitative Comparative Analysis (QCA) [@ragin_1987] and Coincidence Analysis (CNA) [@baumgartner_2009]. For QCA, several software packages exist, each with different functionality [@cronqvist_2019, @dusa_2021, @ragin_davey_2019, @thiem_2018]. However, since there is no harmonization of procedures in QCA, different software packages often generate very different results despite identical data input and parameter settings [@thiem_dusa_2013]. For CNA, the `cna` package is available [@ambuhl_baumgartner_2021].
 
[…]
