+++ paper/paper.md
@@ -41,12 +41,12 @@ QCA and CNA can analyze single outputs only. Although CNA has been built for ana
 
 # Functionality and design 
 
-Figure <a href="#fig:structure" data-reference-type="ref" data-reference="fig:structure">1</a> shows the internal structure of `CORA`. Two distinct Python packages constitute its foundation. While the `CORA` package is responsible for all processes related to algorithmic optimization, the `LOGIGRAM` package provides graphical functionality for the production of logic diagrams. On the middle layer, these two Python packages are combined in CORA's Colab notebook, which provides the visual interface to the user. This end-user interface represents the top level of `CORA` on which all interactions between the user and software take place. The input the user has to provide consists of a set of CORA-compatible data and the specification of all required modelling parameters. Following the process of algorithmic optimization, a solution is generated. In CORA, such solutions take the form of a set of irredundant systems of two-level Boolean functions. These systems, or any stand-alone function in disjunctive normal form provided independently of CORA, can then also be rendered through `LOGIGRAM` as a logic diagram.
+Figure <a href="#fig:structure" data-reference-type="ref" data-reference="fig:structure">1</a> shows the internal structure of `CORA`. Two distinct Python packages constitute its foundation. While the `CORA` package is responsible for all processes related to algorithmic optimization, the `LOGIGRAM` package provides graphical functionality for the production of logic diagrams. On the middle layer, these two Python packages are combined in CORA's Colab notebook, which provides the visual interface to the user. This end-user interface represents the top level of `CORA` on which all interactions between the user and software take place. The input the user has to provide consists of a set of `CORA`-compatible data and the specification of all required modelling parameters. Following the process of algorithmic optimization, a solution is generated. In `CORA`, such solutions take the form of a set of irredundant systems of two-level Boolean functions. These systems, or any stand-alone function in disjunctive normal form provided independently of `CORA`, can then also be rendered through `LOGIGRAM` as a logic diagram.
 
 ![The structure of CORA.\label{fig:structure}](cora_structure.png)
 
 
-The user interface of CORA's Colab notebook is shown in Figure <a href="#fig:notebook" data-reference-type="ref" data-reference="fig:main">2</a>. It guides users through the analysis in nine steps, the last two of which are optional: (1) the initialization of the framework and (2) default settings, the (3) choice and (4) import of data, the (5) specification of the inputs and outputs, the (6) setting of search parameters and thresholds for data fit statistics, (7) the computation of the solution, (8) the initialization of `CORA`'s visualization module and, finally, (9) the drawing and export of logic diagrams. In QCA and CNA, the existence of multiple models that fit some set of data equally well is called model ambiguity [@baumgartner_thiem_2017]. The analogue of model ambiguitiy in CORA is called "system ambiguity". A further commonality with CNA, but difference to QCA, is that CORA does not permit the manipulation of don't care terms. So-called "conservative" or "intermediate" solutions, which produce unacceptably high rates of false positives [@baumgartner_thiem_2020, @thiem_2022b], are not offered in `CORA`. In contrast to both QCA and CNA, CORA does not allow the use of fuzzy variables because of the numerous inferential problems associated with such variables. However, `CORA` comes with full support for multi-value inputs and outputs.     
[…]
