+++ paper/paper.bib
@@ -17,7 +17,7 @@
 @article{baumgartner_thiem_2017,
    author = {Baumgartner, M. and Thiem, A.},
    year = {2017},
-   title = {Model Ambiguities in Configurational Comparative Research},
+   title = {Model ambiguities in configurational comparative research},
    journal = {Sociological Methods \& Research},
    volume = {46},
    number = {4},
@@ -27,7 +27,7 @@
 @article{baumgartner_thiem_2020,
    author = {Baumgartner, M. and Thiem, A.},
    year = {2020},
-   title = {Often Trusted but Never (Properly) Tested: Evaluating Qualitative Comparative Analysis},
+   title = {Often trusted but never (properly) tested: Evaluating {Qualitative Comparative Analysis}},
    journal = {Sociological Methods \& Research},
    volume = {49},
    number = {2},
@@ -113,8 +113,8 @@
 @incollection{thiem_2022a,
    author = {Thiem, A.},
    year = {2022},
-   title = {Qualitative Comparative Analysis (QCA)},
-   booktitle = {Handbook of Research Methods in International Relations},
+   title = {{Qualitative Comparative Analysis (QCA)}},
+   booktitle = {Handbook of research methods in {International Relations}},
    editor = {Huddleston, R. Joseph and Jamieson, Tom and James, Patrick},
    publisher = {Edward Elgar},
    address = {Cheltenham},
@@ -125,12 +125,22 @@
 @article{thiem_2022b,
    author = {Thiem, A.},
    year = {2022},
-   title = {Beyond the Facts: Limited Empirical Diversity and Causal Inference in Qualitative Comparative Analysis},
+   title = {Beyond the facts: Limited empirical diversity and causal inference in {Qualitative Comparative Analysis}},
    journal = {Sociological Methods \& Research},
    volume = {51},
    number = {2},
    pages = {527-540},
    doi = {10.1177/0049124119882463}
+}
+@article{thiem_dusa_2013,
+   author = {Thiem, A. and Du\c{s}a, A.},
+   year = {2013},
+   title = {Boolean minimization in social science research: A review of current software for {Qualitative Comparative Analysis (QCA)}},
+   journal = {Social Science Computer Review},
+   volume = {31},
+   number = {4},
+   pages = {505-521},
+   doi = {10.1177/0894439313478999}
 }
 @article{thiem_et_al_2020,
    author = {Thiem, A. and Mkrtchyan, L. and Haesebrouck, T. and Sanchez, D.},
@@ -148,14 +158,4 @@
    title = {{Combinational Regularity Analysis (CORA)} - A new method for uncovering complex causation in medical and health research},
    journal = {BMC Medical Research Methodology},
    doi = {10.1186/s12874-022-01800-9}
-}
-@article{thiem_dusa_2013,
-   author = {Thiem, A. and Du\c{s}a, A.},
-   year = {2013},
-   title = {Boolean minimization in social science research: A review of current software for {Qualitative Comparative Analysis (QCA)}},
-   journal = {Social Science Computer Review},
-   volume = {31},
-   number = {4},
-   pages = {505-521},
-   doi = {10.1177/0894439313478999}
 }
