+++ paper/paper.bib
@@ -1,6 +1,6 @@
-@book{ambuhl_baumgartner_2021,
+@book{ambuhl_baumgartner_2022,
    author = {Ambuehl, M. and Baumgartner, M.},
-   year = {2020},
+   year = {2022},
    title = {{cna}: Causal modeling with {Coincidence Analysis}, version 3.4.0},
    publisher = {http://cran.r-project.org/package=cna}
 }
@@ -27,7 +27,7 @@
 @article{baumgartner_thiem_2020,
    author = {Baumgartner, M. and Thiem, A.},
    year = {2020},
-   title = {Often trusted but never (properly) tested: Evaluating {Qualitative Comparative Analysis}},
+   title = {Often trusted but never (properly) tested: Evaluating {Qualitative} {Comparative} {Analysis}},
    journal = {Sociological Methods \& Research},
    volume = {49},
    number = {2},
@@ -39,7 +39,7 @@
    year = {2019},
    title = {Tosmana: Tool for small-n analysis, version 1.61 [computer program]},
    publisher = {University of Trier},
-   address = {Trier},
+   address = {Trier}
 }
 @article{delorenzo_schmidt_2018,
    author = {de Lorenzo, V. and Schmidt, M.},
@@ -50,11 +50,11 @@
    pages = {170-180},
    doi = {10.1016/j.nbt.2017.05.001}
 }
-@book{dusa_2021,
+@book{dusa_2022,
    author = {Dusa, A.},
-   year = {2021},
-   title = {{QCA}: {Qualitative Comparative Analysis}, {R} package version 3.11},
-   publisher = {https://cran.r-project.org/web/packages/QCA/},
+   year = {2022},
+   title = {{QCA}: {Qualitative Comparative Analysis}, {R} package version 3.17},
+   publisher = {https://cran.r-project.org/package=QCA}
 }
 @book{lewin_protheroe_1992,
    author = {Lewin, D. and Protheroe, D.},
@@ -79,20 +79,20 @@
    title = {Causality: Models, reasoning, and inference},
    publisher = {Cambridge University Press},
    address = {Cambridge},
-   edition = {2nd},
+   edition = {2nd}
 }
 @book{ragin_1987,
    author = {Ragin, C.C.},
    year = {1987},
    title = {The comparative method: Moving beyond qualitative and quantitative strategies},
    publisher = {University of California Press},
-   address = {Berkeley},
+   address = {Berkeley}
 }
 @book{ragin_davey_2019,
    author = {Ragin, C.C. and Davey, S.},
    year = {2019},
    title = {{fs/QCA}: {fuzzy-set/Qualitative Comparative Analysis}, version 3.1b [Computer Program]},
-   publisher = {http://www.socsci.uci.edu/~cragin/fsQCA/software.shtml},
+   publisher = {http://www.socsci.uci.edu/~cragin/fsQCA/software.shtml}
 }
 @article{suls_green_2019,
    author = {Suls, J. and Green, P.A.},
@@ -102,18 +102,18 @@
    volume = {38},
    number = {9},
    pages = {769-771},
-   doi = {10.1037/hea0000783},
+   doi = {10.1037/hea0000783}
 }
 @book{thiem_2018,
    author = {Thiem, A.},
    year = {2018},
    title = {{QCApro}: Advanced functionality for performing and evaluating {Qualitative Comparative Analysis}, {R} package version 1.1-2},
-   publisher = {http://cran.r-project.org/package=QCApro},
+   publisher = {http://cran.r-project.org/package=QCApro}
 }
 @incollection{thiem_2022a,
    author = {Thiem, A.},
    year = {2022},
-   title = {{Qualitative Comparative Analysis (QCA)}},
[…]
