+++ paper/paper.md
@@ -33,9 +33,9 @@ Combinational Regularity Analysis (CORA) [@thiem_et_al_forth] is a new member of
 
 # Statement of need
 
-Modern CCMs can infer regularity-theoretic causal structures from categorical data. The most sophisticated of these structures are so-called "INUS" structures. "INUS" refers to an *i*nsufficient but *n*on-redundant part of a condition that is *u*nnecessary but *s*ufficient for some outcome [@mackie_1965]. INUS structures are syntactically codified as two-level Boolean functions, such as $a\cdot b' + c' \Leftrightarrow z$, where $a$, $b$, $c$ and $z$ are Boolean literals, $\text{"} \cdot \text{"}$ represents the Boolean multiplication operator, $\text{"} + \text{"}$ the Boolean addition operator, $\text{" } ' \text{ "}$ the Boolean negation operator, and $\text{"}\Leftrightarrow\text{"}$ the Boolean implication operator. Branches of Boolean algebra include, most importantly, set theory, propositional logic and switching circuit theory [@lewin_protheroe_1992].
+Modern CCMs can infer regularity-theoretic causal structures from categorical data. The most sophisticated of these structures are so-called "INUS" structures. "INUS" refers to an **i**nsufficient but **n**on-redundant part of a condition that is **u**nnecessary but **s**ufficient for some outcome [@mackie_1965]. INUS structures are syntactically codified as two-level Boolean functions, such as $a\cdot b' + c' \Leftrightarrow z$, where $a$, $b$, $c$ and $z$ are Boolean literals, $\text{"} \cdot \text{"}$ represents the Boolean multiplication operator, $\text{"} + \text{"}$ the Boolean addition operator, $\text{" } ' \text{ "}$ the Boolean negation operator, and $\text{"}\Leftrightarrow\text{"}$ the Boolean implication operator. Branches of Boolean algebra include, most importantly, set theory, propositional logic and switching circuit theory [@lewin_protheroe_1992].
 
-So far, the two most sophisticated CCMs have been Qualitative Comparative Analysis (QCA) [@ragin_1987] and Coincidence Analysis (CNA) [@baumgartner_2009]. For QCA, several software packages exist, each with different functionality [@cronqvist_2019, @dusa_2021, @ragin_davey_2019, @thiem_2018]. However, since there is no harmonization of procedures in QCA, different software packages often generate very different results despite identical data input and parameter settings [@thiem_dusa_2013]. For CNA, the `cna` package is available [@ambuhl_baumgartner_2021].
+So far, the two most sophisticated CCMs have been Qualitative Comparative Analysis (QCA) [@ragin_1987] and Coincidence Analysis (CNA) [@baumgartner_2009]. For QCA, several software packages exist, each with different functionality [@cronqvist_2019; @dusa_2022; @ragin_davey_2019; @thiem_2018]. However, since there is no harmonization of procedures in QCA, different software packages often generate very different results despite identical data input and parameter settings [@thiem_dusa_2013]. For CNA, the `cna` package is available [@ambuhl_baumgartner_2022].
 
 QCA and CNA can analyze single outputs only. Although CNA has been built for analyses with multiple outputs, its algorithm has to process each output one after another. Many scientific research problems, however, require the simultaneous analysis of multiple effects, such as analyses of multimorbidity [@suls_green_2019]. CORA offers a solution. It generalizes QCA and CNA by allowing the simultaneous analyses of multiple outputs, and can therefore discover more complex INUS structures. To this end, CORA implements adaptations of optimization algorithms originally developed for the design of multi-output switching circuits. In addition, CORA offers a systematic possibility to mine configurational data. Last, but not least, with CORA, researchers can draw on logic diagrams to communicate their results. The software tool `CORA` implements the eponymous method and all aforementioned features.
 
[…]
