### 2022-12-14 #4998 @cprudhom
@vissarion I'm afraid I'm in the same situation. In particular, I cannot judge the functionnality, the purpose of the software or be relevant in comparison with other packages commonly used in this field of research.
