### 2022-12-15 #5019 @editorialbot
[REVIEW]: CORA: A Python Package for Combinational Regularity Analysis
**Submitting author:** @ZuzanaSebb (<a href="http://orcid.org/0000-0002-4669-5659">Zuzana Sebechlebská</a>)
**Repository:** https://github.com/PoliUniLu/cora
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
