+++ README.md
@@ -23,7 +23,7 @@ pip install git+https://github.com/PoliUniLu/cora.git
 
 To open CORA in Google Colab, click the button below:
 
-[![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/gist/ZuzanaSebb/87499b8048d4839e51746ebc319c604d/visualisation-1-0-5.ipynb?authuser=1)
+[![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/gist/ZuzanaSebb/d2fb79f9466af1e74b2de7047d258388/cora-1-0-5.ipynb?authuser=1)
 
 ## Citation Info
 
