### 2022-12-19 #5019 @AlrikThiem
> Hi @ZuzanaSebb the editorial bot just found an invalid DOI in your paper. Please fix this issue.

Hello @vissarion, the DOI is from a paper that has been accepted and is currently in print, but not yet officially published. I will remove the DOI, but leave the forthcoming article in because it's the main intro article to the software. I hope that's ok.
