+++ paper/paper.bib
@@ -156,6 +156,5 @@
    author = {Thiem, A. and Mkrtchyan, L. and Sebechlebská, Z.},
    year = {forthcoming},
    title = {{Combinational Regularity Analysis (CORA)} - A new method for uncovering complex causation in medical and health research},
-   journal = {BMC Medical Research Methodology},
-   doi = {10.1186/s12874-022-01800-9}
+   journal = {BMC Medical Research Methodology}
 }
