+++ paper/paper.bib
@@ -152,9 +152,13 @@
    pages = {e0233625},
    doi = {10.1371/journal.pone.0233625}
 }
-@article{thiem_et_al_forth,
-   author = {Thiem, A. and Mkrtchyan, L. and Sebechlebská, Z.},
-   year = {forthcoming},
-   title = {{Combinational Regularity Analysis (CORA)} - A new method for uncovering complex causation in medical and health research},
-   journal = {BMC Medical Research Methodology}
+@article{thiem_et_al_2022,
+   author = {Thiem, Alrik and Mkrtchyan, Lusine and Sebechlebská, Zuzana},
+   year = {2022},
+   title = {Combinational Regularity Analysis (CORA) — a new method for uncovering complex causation in medical and health research},
+   journal = {BMC Medical Research Methodology},
+   volume = {22},
+   number = {1},
+   pages = {333},
+   doi = {10.1186/s12874-022-01800-9}
 }
