### 2023-01-07 #5019 @Naeemkh
Hi @ZuzanaSebb, I could not find the license that you used for the package under [OSI approved](https://opensource.org/licenses/alphabetical) software license. @vissarion, I am not sure if we need to examine the license case by case. Please let me know.
