+++ README.md
@@ -36,9 +36,7 @@ Sebechlebská, Zuzana, Lusine Mkrtchyan and Alrik Thiem. 2022. CORA: A Python pa
 
 ## Copyright
 
-CORA is licensed under a CC BY-NC-SA 4.0 license. This license allows reusers to distribute, remix, adapt, and build upon the material in any medium or format for noncommercial purposes only, and only so long as attribution is given to the creator. If you remix, adapt, or build upon the material, you must license the modified material under identical terms. 
-
-[![Open In License](https://mirrors.creativecommons.org/presskit/buttons/88x31/svg/by-nc-sa.svg)](https://creativecommons.org/licenses/by-nc-sa/4.0/legalcode)
+CORA is licensed under a GNU GPLv3. 
 
 
 ## References
