### 2023-01-11 #5019 @ZuzanaSebb
Hi @vissarion, we changed the license to GNU GPLv3.
