### 2023-01-13 b7023cb
fix in Colab- multi output
