### 2023-01-16 PoliUniLu/cora#11 @Naeemkh
Running examples on a local system
Thanks for developing the package.
Please add more details on how to run the package on a local system. A step-by-step procedure (`Getting started` or `Tutorial` sections) would be great.
In the current version, without having Google Colab, it is not clear how to use the package.

Related to the following submission: https://github.com/openjournals/joss-reviews/issues/5019
