### 2023-01-16 PoliUniLu/cora#12 @Naeemkh
Contribution guide for open source community
Please add a section on contributions (e.g., `How to contribute?`) for the open-source community.

Related to this submission: https://github.com/openjournals/joss-reviews/issues/5019
