### 2023-01-16 #5019 @Naeemkh
Hi @vissarion, please add details on how to use the package on a local system (at least one running example). There are some codes on Google Colab that I could not find in your submitted GitHub repository. Please see issue [#11](https://github.com/PoliUniLu/cora/issues/11) Thanks.
