### 2023-02-03 7ab42d4
len of the solution test
