### 2023-02-06 efcb696
typo correction in data mining func
