### 2023-02-06 ca4a4c7
change of the names of the input parameters in datamining
