### 2023-02-06 #5019 @ZuzanaSebb
Hi @Naeemkh, we've just added Jupyter notebook (paper/cora_package.ipynb) demonstrating usage of the package.
