### 2023-02-13 #5019 @Naeemkh
Hi @ZuzanaSebb, thanks for submitting the example. Please address the following comments:

- The wording in the paper suggests that Logigram is part of the CORA package; however, it is a standalone package. This can falsely imply that the Logigram has been reviewed through this revision process, which is not the case. Please change the wording, or include the code.
- The package needs a very user-friendly getting started page/documentation and example. Please add steps for setting up the environment (e.g., python version and other dependencies). Also, please bring the example out of the paper folder, preferably in pure python code (not a Jupyter file). The example should be part of the package, not a case to satisfy the reviewer. If the Google Colab document is an important part of the package, why is it not included in the package and is hosted in a random [gist](https://gist.github.com/ZuzanaSebb/223d0ff8817300fab2a038c99d6e5844#file-cora-1-0-6-ipynb)?
- Please add guidance on how the open-source community can contribute to the project. (In this section, please also mention what styling guide you follow in the package.)
