### 2023-02-16 #5019 @ZuzanaSebb
Hi @Naeemkh,

- Logigram is a standalone package, we prefer to keep it in a separate repository. We changed the wording in the paper.
- We added more explicit examples in a dedicated folder (pure python and a  Jupyter notebook). The dependencies for
the package are specified in a pyproject.toml (using Poetry), so they all can be installed automatically by running the pip command listed in the README.md file. The Colab notebook was created for our users not familiar with python.
- We updated the readme file with a short note about the contributions.
