### 2023-02-17 #5019 @vissarion
@Naeemkh please let us know if your raised issues are addressed or you need a second round.
@ZuzanaSebb there are still two issues related with @Naeemkh review in the repository https://github.com/PoliUniLu/cora/issues
@Arrondissement5etDemi your review is missing a checkbox on functionality. Is this intentional? Are you waiting for some input from the authors?
