### 2023-02-18 PoliUniLu/cora#13 @Naeemkh
Inconsistency in naming conventions
Please choose a naming convention and use it throughout the codebase, for example the following function and class names are not consistent with the code.
- `_prepareRows` in the following line: https://github.com/PoliUniLu/cora/blob/6efe5ca630c53f5e2bc2b9f339b5118b5a361da9/cora/prime_implicants.py#L630
- `On_off_reduction_matrix` in this line:  https://github.com/PoliUniLu/cora/blob/6efe5ca630c53f5e2bc2b9f339b5118b5a361da9/cora/on_off_alg.py#L80

This issue is related to this submission: https://github.com/openjournals/joss-reviews/issues/5019
