### 2023-02-18 PoliUniLu/cora#14 @Naeemkh
API Documentation
API references are hard to find. A user will need help understanding different input parameters and other details and notes.

- For example in the provided examples [here](https://github.com/PoliUniLu/cora/blob/6efe5ca630c53f5e2bc2b9f339b5118b5a361da9/examples/example1.py) you are using `cora.OptimizationContext`. How can a user get more information about it? Is there any API reference page to support it? Using `help(cora.OptimizationContext)` generates very inconclusive results.  You have added documentation to that class, but you do not leverage it in the package.  Automating the process of document generation is crucial to ensure that documentation remains up-to-date. I noticed that in the `prime_implicants.OptimizationContext` [class](https://github.com/PoliUniLu/cora/blob/6efe5ca630c53f5e2bc2b9f339b5118b5a361da9/cora/prime_implicants.py#L359), some parameters are not properly documented, such as `temporal_labels`. It is important to have a systematic way of generating documentation to avoid such issues.

This issue is related to this submission: https://github.com/openjournals/joss-reviews/issues/5019
