### 2023-02-18 PoliUniLu/cora#15 @Naeemkh
Developer's guide
Thanks for adding the developer's guide.
- Please consider adding Continuous Integration into the package (using Github actions), **or** please add good documentation on how to test the package.
- In the current version, it is not clear what percentage of the package is tested.
- It is also hard to test the package locally.
- For example, what is `parametrized` in the following line: https://github.com/PoliUniLu/cora/blob/6efe5ca630c53f5e2bc2b9f339b5118b5a361da9/tests/test_implicants.py#L3
- Where is the data in this unittest:
https://github.com/PoliUniLu/cora/blob/6efe5ca630c53f5e2bc2b9f339b5118b5a361da9/tests/test_implicants.py#L14

Please add documentation for the developers. If the developing environment differs from the user environment, please also discuss it in the package.

This issue is related to this submission: https://github.com/openjournals/joss-reviews/issues/5019
