### 2023-02-18 #5019 @Naeemkh
Thank you for addressing my comments. I believe that the CORA package is a powerful tool for Combinational Regulatory Analyses and has great potential to contribute to the open-source community. The code is well-developed, but some minor issues need to be addressed in order for it to be acceptable for publication. Please refer to the following submitted issues for detailed requirements.

- I suggest that the authors have other team or lab members use and test the package and encourage them to see if they have all the required ingredients to contribute to its development. While I am happy to provide feedback in this review thread, it is ultimately up to the developers to take proactive steps to address any issues and improve the package. Delaying action on their part will only prolong the review process.
- Please make sure to open a new git branch and submit a pull request to yourself so that reviewers can see all the changes made from one revision to another.
- Regarding the LOGIGRAM package, please treat it as a third-party dependency and do not discuss it in the software paper. For example, in line 86 of the software paper, you included extra future functionality to the LOGIGRAM package.
- To make the package more user-friendly, I suggest adding a usage section to the README file and providing a simple example. This will help users understand how to use the package and get started with their analyses.
-  I recommend removing any commented-out lines of code. This will make the code cleaner and easier to read, while still allowing you to retrieve that code line by checking out old commits if necessary.

Please see the following submitted issues:
- https://github.com/PoliUniLu/cora/issues/13
- https://github.com/PoliUniLu/cora/issues/14
- https://github.com/PoliUniLu/cora/issues/15
