+++ cora/prime_implicants.py
@@ -23,7 +23,6 @@ COLUMN_LABELS = list(string.ascii_uppercase) + ["AA", "BB", "CC", "DD",
                                                 "EE", "FF"]
 OUTPUT_PATTERN = re.compile("^([a-zA-Z0-9]+)\{([0-9]+(,[0-9]+)*)\}$")
 REGULAR_OUTPUT = re.compile("^([a-zA-Z0-9]+)")
-TEMPORAL_COL_PATTERN = re.compile("^([a-zA-Z0-9-\_]+)\{([0-9]+(,[0-9]+)*)\}$")
 
 
 def concatenate_strings(arr):
@@ -73,17 +72,18 @@ def is_minterm_subset(m1, m2):
 
 # Funtion which groups minterms according to the number of nonzero digets
 
-# Paramters: truth table, 
+# Paramters: table - the truth table
 #            column_number - defines the number of possible nonzero digits
 #            cares - indexes which correspond to original minterms
-#            outputcolumns - binary output columns from the original data
-#            multi_output - a bolean variable
-# Output: an array res of objects from classes Multi_value_item or
-#         Multiple_output_minterm, the i-th element of an array consist of 
-#         objects whose minterms have exactly i nonzero digits
+#            output_columns - binary output columns from the original data
+#            multi_output - a boolean variable
+# Output: (a) an array res of objects from classes Multi_value_item or
+#             Multiple_output_minterm
+#         (b) the i-th element of an array consisting of
+#              objects whose minterms have exactly i nonzero digits
 
 # Definition: a minterm of n variables is a product of the variables
-# 	      in which each appears exactly once in true or complemented form. 
+# 	      in which each appears exactly once in true or complemented form.
 
 def create_groups(table, column_number, cares, outputcolumns, multi_output):
     res = []
@@ -93,7 +93,7 @@ def create_groups(table, column_number, cares, outputcolumns, multi_output):
     if not multi_output:
         for row in table:
             non_zeros = count_non_zeros(row)
-            res[non_zeros].append(Multi_value_minterm(row, set([index])
+            res[non_zeros].append(MultiValueMinterm(row, set([index])
             if index in cares else set()))
             index = index + 1
         return res
@@ -111,7 +111,7 @@ def create_groups(table, column_number, cares, outputcolumns, multi_output):
             if len(tag) == 0:
                 index = index + 1
                 continue
-            res[non_zeros].append(Multiple_output_minterm(row, set([index])
+            res[non_zeros].append(MultipleOutputMinterm(row, set([index])
             if index in cares else set(), tag))
             index = index + 1
         return res
@@ -119,19 +119,19 @@ def create_groups(table, column_number, cares, outputcolumns, multi_output):
 
 # Function preforming 1 step elimination of 2 minterms
 
-# Function eliminates minterms first in adjacting groups, second in
-# between the group. In both cases minterms differ only in one sigle digit.
-# In addition, the function collects implicants, which can't be further 
+# Function eliminates minterms first in adjacent groups, second in
+# between the group. In both cases minterms differ in only one sigle digit.
+# In addition, the function collects implicants, which can't be further
 # eliminated to an array- final implicants.
 
 # Parameters: groups - an array returned by fun create_groups
 #             n - the number of rows in the truth table
 #             multi_output - a boolean variable
 
-# Output: dictionary, which contains 
-# 	  (a) new groups of minterms after the elimination, 
-#         (b) implicants   
-#         (c) information if any reduction was performed in this single 
+# Output: dictionary, which contains
+# 	  (a) new groups of minterms after the elimination
+#         (b) implicants
+#         (c) information if any reduction was performed in this single
 #             elimination
 
[…]
