+++ cora/data_mining_cora.py
@@ -1,13 +1,10 @@
 import itertools
 import pandas as pd
 import cora
-import re
-
-TEMPORAL_COL_PATTERN = re.compile("^([a-zA-Z0-9-\_]+)\{([0-9]+(,[0-9]+)*)\}$")                                              
 
 
 class TupleResult:
-    
+
     def __init__(self,
                  combination,
                  out_len,
@@ -15,211 +12,146 @@ class TupleResult:
                  inc_score,
                  cov_score,
                  score):
-        
-        self.combination=combination
-        if(out_len ==1 and len(irrendudant_systems) == 1 and (str(
-                irrendudant_systems[0].system[0].implicant)=='1')):
+
+        self.combination = combination
+        if (out_len == 1 and len(irrendudant_systems) == 1 and (str(
+                irrendudant_systems[0].system[0].implicant) == '1')):
             self.nr_irr_systems = 0
             self.inc_score = 0
             self.cov_score = 0
-            self.score=score = 0
-        else:     
+            self.score = score = 0
+        else:
             self.nr_irr_systems = len(irrendudant_systems)
             self.inc_score = inc_score
             self.cov_score = cov_score
             self.score = score
 
-#valid_comb: ((1, 'PUBLIC'), (2, 'ELITE')) , ['E_BEFORE_A{2,3}']
-
-def valid_combination(combination,temp_cols):
-    #combination_cols  = set(x[1] for x in combination) 
-    #temp_cols_string = {m.group(1) 
-     #                   for m in [TEMPORAL_COL_PATTERN.match(i)
-     #                   for i in temp_cols]}
-    if len(temp_cols)<1: 
-       return True
- 
-    else:
-        cols_indexes = set(x[0] for x in combination)
-        tmp_t_labels = {m.group(2) 
-                        for m in [TEMPORAL_COL_PATTERN.match(i)
-                        for i in temp_cols]}
-        temp_index_tmp = [x.split(",") for x in tmp_t_labels]
-        temp_index = {int(x) for y in temp_index_tmp for x in y}
-        
-        return temp_index.issubset(cols_indexes)
- 
-
-    
-    
-"""
-
-Parameters
-
-----------
-
-data : dataframe
-
-output_labels : an array of strings
-    The names of the outcome columns from the data frame.
-    
-len_of_tuple : int
-    Number indicating how many variables from the original data 
-    are used in the computation.
-
-case_col : string
-    The name of the column from the data frame containing the case ids.
-
-n_cut : int
-    The minimum number of cases under which a truth table row is declared as a
-    remainder
-        
-inc_score1 : float
-	The minimum sufficiency inclusion score for an output function value of "1"
-    
-inc_score2 : float
-	The maximum sufficiency inclusion score for an output function value of "0"
-
-U : int
-    The U number is either 0 or 1.
-
-algorithm : string
-            ON-DC or ON-OFF
-        
-
-Returns
--------
-
-result : dataframe
-    
-    
-"""   
 
 def data_mining(data,
                 output_labels,
                 len_of_tuple,
                 case_col=None,
-                temp_cols=None,
                 n_cut=1,
                 inc_score1=1,
                 inc_score2=None,
-                Uvalue=None,
+                U=None,
                 algorithm="ON-DC"):
+    '''
+
+    Function performing data mining on a subset of input variables.
+    Parameters
+    ----------
+    data : dataframe
+
+    output_labels : an array of strings
+    The names of the outcome columns from the data frame.
+
+    len_of_tuple : int
+    Number indicating how many input variables from the original data
+    are used in the computation.
+
+    case_col : string
+    The name of the column from the data frame containing the case ids.
+
+    n_cut : int
[…]
