### 2023-03-03 6bb75ae
Update prime_implicants.py
