+++ .github/workflows/python-package.yml
@@ -28,8 +28,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install flake8 pytest
-        poetry.lock
-        if [ -f poetry.lock ]; then pip install -r poetry.lock; fi
+        pip install -r pyproject.toml
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
