+++ .github/workflows/python-package.yml
@@ -28,7 +28,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install flake8 pytest
-        pip install -r pyproject.toml
+        pip install -r poetry.lock
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
