### 2023-03-03 173cba8
fix in test pi_chart
