### 2023-03-03 106c008
poetry run pytest
test poetry

test githubworkflow

poetry changes

change in poetry

change in poetry

poetry fix

test datasets

fix in data type of inc.score

status badge

coverage.yml

fix in syntax coverage.yml

syntax fix in coverage.yml

fix in codecov syntax

new coverage.yml

fix in coverage.yml

fix in coverage.yml

upload to codecov

changes in coverage.yml

changes in cov.yml

detailed cov. report

workflow github actions, tests + coverage, fixes in the tests
