+++ .github/workflows/coverage.yml
@@ -1,0 +1,46 @@
+name: Coverage
+
+on:
+  push:
+    branches: [ "master" ]
+  pull_request:
+    branches: [ "master" ]
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Python 
+      uses: actions/setup-python@v3
+      with:
+          python-version: 3.8
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        python -m pip install coverage
+        python -m pip install pytest
+        python -m pip install poetry
+    - name: Set poetry config
+      run: python -m poetry config virtualenvs.create false
+    - name: Install dependencies
+      run: python -m poetry install
+    
+    - name: Test with coverage
+      run: |
+          poetry run coverage run -m pytest tests/
+          poetry run coverage report -i
+          poetry run coverage xml -i
+    - name: Upload coverage to Codecov
+      uses: codecov/codecov-action@v3
+      with:
+          fail_ci_if_error: true
+          token: ${{ secrets.CODECOV_TOKEN }}
+         
+ 
+
+
+
+
+
