+++ .github/workflows/python-package.yml
@@ -28,8 +28,11 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install flake8 pytest
-        pip install poetry==1.*
-        poetry install
+        python -m pip install poetry
+    - name: Set poetry config
+      run: python -m poetry config virtualenvs.create false
+    - name: Install dependencies
+      run: python -m poetry install
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
@@ -37,5 +40,5 @@ jobs:
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
-      run: |
-        pytest
+      run: poetry run pytest tests/
+    
