+++ cora/__init__.py
@@ -1,4 +1,3 @@
 from .prime_implicants import OptimizationContext
 from .petric import find_irredundant_sums
-from .preprocess_data import preprocess_data
 from .data_mining_cora import data_mining
