### 2023-03-10 a8a3d42
removal of the temporal variable + data mining tests
