+++ cora/prime_implicants.py
@@ -2,7 +2,6 @@
 This module contains the implementation of Boolean minimization algorithms.
 
 """
-
 import numpy as np
 import pandas as pd
 import itertools
@@ -23,7 +22,6 @@ COLUMN_LABELS = list(string.ascii_uppercase) + ["AA", "BB", "CC", "DD",
                                                 "EE", "FF"]
 OUTPUT_PATTERN = re.compile("^([a-zA-Z0-9]+)\{([0-9]+(,[0-9]+)*)\}$")
 REGULAR_OUTPUT = re.compile("^([a-zA-Z0-9]+)")
-TEMPORAL_COL_PATTERN = re.compile("^([a-zA-Z0-9-\_]+)\{([0-9]+(,[0-9]+)*)\}$")
 
 
 def concatenate_strings(arr):
@@ -250,23 +248,6 @@ def eliminate_minterms(table, elements, levels, multi_output):
     return [x for we, x in zip(was_eliminated, decomposed) if not we]
 
 
-def eliminate_useless_temp_implicants(impl_str, temporal_labels, input_labels):
-    implicant_parts = [i.lower() for i in impl_str.split('*')]
-    tmp_t_labels = {m.group(1): m.group(2)
-                    for m in [TEMPORAL_COL_PATTERN.match(i)
-                              for i in temporal_labels]}
-
-    t_labels = tmp_t_labels.keys()
-    for label in t_labels:
-        if label.lower() in implicant_parts:
-            indx = [int(x) - 1 for x in tmp_t_labels[label].split(",")]
-            for possible_literal in [input_labels[i] for i in indx]:
-                if possible_literal.lower() not in implicant_parts:
-                    return False
-
-    return True
-
-
 def find_index2(arr, x):
     return np.where((arr == x).all(axis=1))[0]
 
@@ -345,13 +326,7 @@ U : int
 rename_columns: boolean 
     If true, the column names are renamed as single letters in alphabetic
     order.
-    
-generateMissing: boolean
-    If false, the conservative solution is calculated.
-    Rows corresponding to the 'don't care' combinations are not
-    taken into the calculation.
-
-        
+  
 
 """
 
@@ -363,13 +338,11 @@ class OptimizationContext:
                  output_labels,
                  input_labels=None,
                  case_col=None,
-                 temporal_labels=None,
                  n_cut=1,
                  inc_score1=1,
                  inc_score2=None,
                  U=None,
                  rename_columns=False,
-                 generateMissing=True,
                  algorithm="ON-DC"
                  ):
         self.data = data.copy()
@@ -383,8 +356,6 @@ class OptimizationContext:
         self.rename_columns = rename_columns
         self.case_col = case_col
         self.output_labels = output_labels
-        self.temporal_labels = temporal_labels
-        self.generate_missing = generateMissing
         self.prime_implicants = None
         self.irredundat_sums = None
         self.irredundant_systems = None
@@ -407,88 +378,16 @@ class OptimizationContext:
         class InvalidDataException(Exception):
             pass
 
-        # temporal columns
-        # auxiliary, indicator columns
+        if self.input_labels is None:
 
-        if self.temporal_labels is not None:
-            if any(TEMPORAL_COL_PATTERN.match(i) is None for i in
-                   self.temporal_labels):
-                raise RuntimeError("Unsupported temp columns entered!")
-
-            # valid data
-            tmp_t_labels = {m.group(1): m.group(2)
-                            for m in [TEMPORAL_COL_PATTERN.match(i)
-                                      for i in self.temporal_labels]}
-
-            t_labels = tmp_t_labels.keys()
-            if (not all(self.data[t_labels].apply(
-                    lambda row_series: all((pd.isna(x)
-                                            or int(x) == 1
-                                            or int(x) == 0)
[…]
