+++ cora/prime_implicants.py
@@ -91,7 +91,7 @@ def create_groups(table, column_number, cares, outputcolumns, multi_output):
     if not multi_output:
         for row in table:
             non_zeros = count_non_zeros(row)
-            res[non_zeros].append(Multi_value_minterm(row, set([index])
+            res[non_zeros].append(MultiValueMinterm(row, set([index])
             if index in cares else set()))
             index = index + 1
         return res
@@ -109,7 +109,7 @@ def create_groups(table, column_number, cares, outputcolumns, multi_output):
             if len(tag) == 0:
                 index = index + 1
                 continue
-            res[non_zeros].append(Multiple_output_minterm(row, set([index])
+            res[non_zeros].append(MultipleOutputMinterm(row, set([index])
             if index in cares else set(), tag))
             index = index + 1
         return res
@@ -435,7 +435,7 @@ class OptimizationContext:
 
         self.validation = True
 
-    def _preprocess_data(self):
+    def preprocess_data(self):
 
         if not self.validation:
             self.data_validation()
@@ -492,7 +492,7 @@ class OptimizationContext:
 
     def get_preprocessed_data(self, raw=False):
         if not self.preprocessing:
-            self._preprocess_data()
+            self.preprocess_data()
         if raw:
             return self.preprocessed_data_raw
         else:
@@ -516,9 +516,9 @@ class OptimizationContext:
 
         return levels
 
-    def _prepareRows(self):
+    def prepareRows(self):
         if not self.preprocessing:
-            self._preprocess_data()
+            self.preprocess_data()
         mask1 = self.preprocessed_data[self.output_labels]
         nr_rows = len(self.preprocessed_data.index)
         n = len(mask1.columns)
@@ -599,7 +599,7 @@ class OptimizationContext:
     def get_prime_implicants_1_DC(self):
 
         if not self.prepare_rows_called:
-            self._prepareRows()
+            self.prepareRows()
         if len(self.table) == 0:
             prime_implicants = tuple()
             return prime_implicants
@@ -632,7 +632,7 @@ class OptimizationContext:
 
         if self.multi_output:
 
-            prime_implicants_fin = tuple(Implicant_multi_output(
+            prime_implicants_fin = tuple(ImplicantMultiOutput(
                 self,
                 minterm_to_str(x[0],
                                self.levels,
@@ -726,7 +726,7 @@ class OptimizationContext:
     def get_prime_implicants_on_off(self):
 
         if not self.preprocessing:
-            self._preprocess_data()
+            self.preprocess_data()
         #  constant outputs
         outputs = self.preprocessed_data[self.output_labels]
 
@@ -760,16 +760,16 @@ class OptimizationContext:
                                                              zip(row, raw_im)),
                                              axis=1)].index])
 
-                    tmp_res.append(Implicant_multi_output(self,
-                                                          minterm_to_str(raw_im,
+                    tmp_res.append(ImplicantMultiOutput(self,
+                                                        minterm_to_str(raw_im,
                                                                          self.levels,
                                                                          self.labels,
                                                                          0,
                                                                          self.multi_output),
-                                                          raw_im,
-                                                          cov,
-                                                          o_tag,
-                                                          self.output_labels))
+                                                        raw_im,
+                                                        cov,
+                                                        o_tag,
+                                                        self.output_labels))
             res = []
             for i in tmp_res:
                 add_to_res = True
@@ -944,7 +944,7 @@ class OptimizationContext:
                                               )
         irredundant_objects = []
         for i, system in enumerate(result):
[…]
