+++ cora/prime_implicants.py
@@ -1,7 +1,7 @@
 """
 This module contains the implementation of Boolean minimization algorithms.
-
 """
+
 import numpy as np
 import pandas as pd
 import itertools
@@ -29,7 +29,7 @@ def concatenate_strings(arr):
 
 
 def inclusion_score(arr):
-    return float(round(sum(arr) / len(arr), 2))
+    return round(sum(arr) / len(arr), 2)
 
 
 def count_non_zeros(minterm):
@@ -69,19 +69,20 @@ def is_minterm_subset(m1, m2):
     return True
 
 
-# Funtion which groups minterms according to the number of nonzero digets
+# Funtion which groups minterms according to the number of non-zero digets
 
-# Paramters: truth table, 
-#            column_number - defines the number of possible nonzero digits
-#            cares - indexes which correspond to original minterms
-#            outputcolumns - binary output columns from the original data
-#            multi_output - a bolean variable
-# Output: an array res of objects from classes Multi_value_item or
-#         Multiple_output_minterm, the i-th element of an array consist of 
-#         objects whose minterms have exactly i nonzero digits
+# Paramters: table - the truth table
+#            column_number - defines the number of possible non-zero digits
+#            cares - indexes of the positive mintermsin in the data
+#            output_columns - binary output columns from the data
+#            multi_output - a boolean variable
+# Output: (a) an array res of objects from classes Multi_value_item or
+#             Multiple_output_minterm
+#         (b) the i-th element of an array consisting of
+#              objects whose minterms have exactly i non-zero digits
 
 # Definition: a minterm of n variables is a product of the variables
-# 	      in which each appears exactly once in true or complemented form. 
+# 	      in which each one appears exactly once in true or complemented form.
 
 def create_groups(table, column_number, cares, outputcolumns, multi_output):
     res = []
@@ -115,21 +116,21 @@ def create_groups(table, column_number, cares, outputcolumns, multi_output):
         return res
 
 
-# Function preforming 1 step elimination of 2 minterms
+# Function performing 1 step elimination of 2 minterms
 
-# Function eliminates minterms first in adjacting groups, second in
-# between the group. In both cases minterms differ only in one sigle digit.
-# In addition, the function collects implicants, which can't be further 
-# eliminated to an array- final implicants.
+# Function eliminates minterms first in adjacent groups, second in
+# between the group. In both cases minterms differ in only one sigle digit.
+# In addition, the function collects implicants, which can't be further
+# eliminated to an array- the prime implicants.
 
 # Parameters: groups - an array returned by fun create_groups
 #             n - the number of rows in the truth table
 #             multi_output - a boolean variable
 
-# Output: dictionary, which contains 
-# 	  (a) new groups of minterms after the elimination, 
-#         (b) implicants   
-#         (c) information if any reduction was performed in this single 
+# Output: dictionary, which contains
+# 	  (a) new groups of minterms after the elimination
+#         (b) implicants
+#         (c) information if any reduction was performed in this single
 #             elimination
 
 def reduction_step(groups, n, multi_output):
@@ -174,7 +175,7 @@ def reduction_step(groups, n, multi_output):
              'reduction': was_any_reduction})
 
 
-# Next set of 3 functions perform an extra elimination step 
+# Next set of 3 functions perform an extra elimination step
 
 
 def extend_with_first(first, arr, multi_output):
@@ -248,6 +249,7 @@ def eliminate_minterms(table, elements, levels, multi_output):
     return [x for we, x in zip(was_eliminated, decomposed) if not we]
 
 
+
 def find_index2(arr, x):
     return np.where((arr == x).all(axis=1))[0]
 
[…]
