+++ paper/paper.md
@@ -41,9 +41,8 @@ QCA and CNA can analyze single outputs only. Although CNA has been built for ana
 
 # Functionality and design 
 
-Figure <a href="#fig:structure" data-reference-type="ref" data-reference="fig:structure">1</a> shows the internal structure of `CORA`. Two distinct Python packages constitute its foundation. While the `CORA` package is responsible for all processes related to algorithmic optimization, the `LOGIGRAM` package provides graphical functionality for the production of logic diagrams.[^1] On the middle layer, these two Python packages are combined in CORA's Colab notebook, which provides the visual interface to the user. This end-user interface represents the top level of `CORA` on which all interactions between the user and software take place. The input the user has to provide consists of a set of `CORA`-compatible data and the specification of all required modelling parameters. Following the process of algorithmic optimization, a solution is generated. In `CORA`, such solutions take the form of a set of irredundant systems of two-level Boolean functions. These systems, or any stand-alone function in disjunctive normal form provided independently of `CORA`, can then also be rendered through `LOGIGRAM` as a logic diagram.
+Figure <a href="#fig:structure" data-reference-type="ref" data-reference="fig:structure">1</a> shows the internal structure of `CORA`. Two distinct Python packages constitute its foundation. While the `CORA` package is responsible for all processes related to algorithmic optimization, the `LOGIGRAM` package provides graphical functionality for the production of logic diagrams. On the middle layer, these two Python packages are combined in CORA's Colab notebook, which provides the visual interface to the user. This end-user interface represents the top level of `CORA` on which all interactions between the user and software take place. The input the user has to provide consists of a set of `CORA`-compatible data and the specification of all required modelling parameters. Following the process of algorithmic optimization, a solution is generated. In `CORA`, such solutions take the form of a set of irredundant systems of two-level Boolean functions. These systems, or any stand-alone function in disjunctive normal form provided independently of `CORA`, can then also be rendered through `LOGIGRAM` as a logic diagram.
 
-[^1]:The `LOGIGRAM` package has not been reviewed as part of the submission of this article to the Journal of Open Source Software.
 
 ![The structure of CORA.\label{fig:structure}](cora_structure.png)
 
