### 2023-03-12 #5019 @ZuzanaSebb
Dear @Naeemkh,
To address your comments:
- We improved the testing process by creating automated actions on GitHub.
- Revisited the code and excluded some parts which were mostly experimental (mainly processing of the temporal data).
- To improve the documentation we incorporated docstrings into the code and added a few examples to the Readme file.
- Readme file contains a short section regarding the development guide.

After further consideration of removing the part of the Logigram package from the paper, we would like to kindly ask you to review the Logigram package as well. It was developed alongside the Cora package as an independent visualization tool which also became one of the important features of the project.
