### 2023-03-16 PoliUniLu/cora#13 @Naeemkh
This issue remains unresolved; however, I will close it for now. Please keep it in mind for consideration during any future refactoring efforts.
