### 2023-03-16 #5019 @Naeemkh
Hi @ZuzanaSebb,

Thank you for addressing my comments. Although the styling convention [issue](https://github.com/PoliUniLu/cora/issues/13) remains unresolved, I have closed it. There are still three outstanding issues:

- Many functions and classes lack API method documentation.
- Some of the unit tests have failed. Please see [issue 17](https://github.com/PoliUniLu/cora/issues/17).
- LOGIGRAM package review

@vissarion, could you please advise us on how to proceed with the LOGIGRAM package issue? To summarize, the paper presents LOGIGRAM as if it is part of CORA or the same project, but the code for LOGIGRAM has not been submitted for review. Thank you.
