+++ README.md
@@ -11,21 +11,21 @@ CORA is a member of the family of Configurational Comparative Methods (CCMs; Thi
 
 In contrast to QCA and CNA, however, CORA is inspired by switching circuit analysis, a subfield of electrical engineering. INUS structures and switching circuits have much in common because propositional logic - the language of INUS causation - and switching algebra - the language of switching circuit analysis - are two conceptually distinct yet operationally equivalent branches of the same underlying Boolean algebra (Lewin and Protheroe 1992). 
 
-CORA is currently the only CCM able to analyze INUS structures that simultaneously feature simple as well as conjunctively complex effects (e.g., *y* **and** **not** *z*, **not** *y* **and** *z*, *y* **and** *z*). In addition, CORA offers a configurational version of Occam's Razor: a data-mining approach to solution building that reduces model ambiguities by keeping the number of required variables for finding a solution at a minimum. Lastly, CORA includes a lean yet powerful visualization module called LOGIGRAM, with which two-level logic diagrams can be produced from any (system of) Boolean function(s) in disjunctive normal form.
+CORA is currently the only CCM able to analyze INUS structures that simultaneously feature simple as well as conjunctively complex effects (e.g., *y* **and** **not** *z*, **not** *y* **and** *z*, *y* **and** *z*), even in multi-value form (Mkrtchyan *et al*. 2023). In addition, CORA offers a configurational version of Occam's Razor: a data-mining approach to solution building that reduces model ambiguities by keeping the number of required variables for finding a solution at a minimum. Lastly, CORA includes a lean yet powerful visualization module called LOGIGRAM, with which two-level logic diagrams can be produced from any (system of) Boolean function(s) in disjunctive normal form (Thiem *et al*. 2023).
 
 ## Installation
 
-CORA requires Python>3.7 and uses Poetry for dependendecy management. Use the package manager [pip](https://pip.pypa.io/en/stable/) to install CORA including all dependencies.
+CORA requires Python>3.7 and uses Poetry for dependendecy management. Use the package manager [pip](https://pip.pypa.io/en/stable/) to install CORA, including all dependencies.
 
 ```bash
 pip install git+https://github.com/PoliUniLu/cora.git
 ```
 
-It's recommended to install the package into dedicated virtual environment.
+It is recommended to install the package into a dedicated virtual environment.
 
 ## Google Colab
 
-To open CORA in Google Colab, click the button below:
+To open CORA with a graphical interface in Google Colab, click the button below:
 
 [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/gist/ZuzanaSebb/5bb6e29772dea02c0c562979f52fd27a/cora-1-0-6.ipynb?authuser=1)
 ## Usage
@@ -109,8 +109,8 @@ Test the code by running
 ```console
 poetry run pytest
 ```
-Pull requests are welcome. Note that although the current codebase doesn't have entirely 
-consistent code style the new code should be PEP-8 compliant.
+Pull requests are welcome. Note that, although the current codebase does not have an entirely 
+consistent code style, the new code should be PEP-8 compliant.
 
 ## References
 
@@ -119,9 +119,11 @@ consistent code style the new code should be PEP-8 compliant.
[…]
