+++ tests/test_multi_value_output.py
@@ -10,7 +10,7 @@ class TestOutputs(unittest.TestCase):
                             columns=['a','b','c','o'])
         context = cora.OptimizationContext(df,['o'])
         with self.assertRaises(Exception):
-            preprocessed_data = context.preprocess_data()
+            preprocessed_data = context._preprocess_data()
     def test_output(self):
         df = pd.DataFrame(
             [[1, 0, 1, 1], [0, 1, 1, 1], [0, 1, 0, 0], [1, 0, 1, 2]],
@@ -19,11 +19,13 @@ class TestOutputs(unittest.TestCase):
         df_test1 = pd.DataFrame(
             [[0, 1, 0, 0], [0, 1, 1, 1],[1, 0, 1, 0]],
             columns=['a', 'b', 'c', 'o'])
+        df_test1.astype({'o': 'int64'})
         assert_frame_equal(df_test1, context1.get_preprocessed_data())
         context2 = cora.OptimizationContext(df, ['o{1,2}'])
         df_test2 = pd.DataFrame(
             [[0, 1, 0, 0], [0, 1, 1, 1],[1, 0, 1, 1]],
             columns=['a', 'b', 'c', 'o'])
+        df_test2.astype({'o': 'int64'})
         assert_frame_equal(df_test2, context2.get_preprocessed_data())
 
 
