+++ cora/__init__.py
@@ -1,4 +1,5 @@
-from .prime_implicants import OptimizationContext
+from .prime_implicants import OptimizationContext, Implicant,\
+    ImplicantMultiOutput,IrredundantSystemsMulti,IrredundantSystem
 from .petric import find_irredundant_sums
 from .data_mining_cora import data_mining
 from .draft_cubes import find_implicants_cubes as experimental_find_implicants_cubes
