+++ cora/prime_implicants.py
@@ -17,22 +17,22 @@ from .multiply import transform_to_raw_implicant
 from .draft_cubes import find_implicants_cubes, transform_to_raw_imp
 from .essential import get_essential_implicants, transform_to_raw_impl, \
     reduce_the_onset
-
+from .boom_1 import best_literals
 COLUMN_LABELS = list(string.ascii_uppercase) + ["AA", "BB", "CC", "DD",
                                                 "EE", "FF"]
 OUTPUT_PATTERN = re.compile("^([a-zA-Z0-9]+)\{([0-9]+(,[0-9]+)*)\}$")
 REGULAR_OUTPUT = re.compile("^([a-zA-Z0-9]+)")
 
 
-def concatenate_strings(arr):
+def _concatenate_strings(arr):
     return ','.join([str(x) for x in arr])
 
 
-def inclusion_score(arr):
+def _inclusion_score(arr):
     return round(sum(arr) / len(arr), 2)
 
 
-def count_non_zeros(minterm):
+def _count_non_zeros(minterm):
     n = len(minterm)
     non_zeros = 0
     for i in range(0, n):
@@ -43,26 +43,26 @@ def count_non_zeros(minterm):
     return non_zeros
 
 
-def vector_to_minterm(v):
+def _vector_to_minterm(v):
     return tuple(frozenset({x}) for x in v)
 
 
-def preprocess_input(table):
-    return [vector_to_minterm(x) for x in table]
+def _preprocess_input(table):
+    return [_vector_to_minterm(x) for x in table]
 
 
-def create_care_translation_dict(cares, labels):
+def _create_care_translation_dict(cares, labels):
     return {k: v for k, v in zip(cares, labels)}
 
 
-def check_element_coverage(row, element):
+def _check_element_coverage(row, element):
     for r, e in zip(row, element[0]):
         if r not in e:
             return False
     return True
 
 
-def is_minterm_subset(m1, m2):
+def _is_minterm_subset(m1, m2):
     for x, y in zip(m1[0], m2[0]):
         if not x.issubset(y):
             return False
@@ -84,14 +84,14 @@ def is_minterm_subset(m1, m2):
 # Definition: a minterm of n variables is a product of the variables
 # 	      in which each one appears exactly once in true or complemented form.
 
-def create_groups(table, column_number, cares, outputcolumns, multi_output):
+def _create_groups(table, column_number, cares, outputcolumns, multi_output):
     res = []
     for i in range(0, column_number + 1):
         res.append([])
     index = 0
     if not multi_output:
         for row in table:
-            non_zeros = count_non_zeros(row)
+            non_zeros = _count_non_zeros(row)
             res[non_zeros].append(MultiValueMinterm(row, set([index])
             if index in cares else set()))
             index = index + 1
@@ -100,7 +100,7 @@ def create_groups(table, column_number, cares, outputcolumns, multi_output):
         nr_of_outputs = len(outputcolumns)
         care_index = 0
         for row in table:
-            non_zeros = count_non_zeros(row)
+            non_zeros = _count_non_zeros(row)
             if index in cares:
                 tag = {i for i in range(1, nr_of_outputs + 1)
                        if outputcolumns[i - 1][care_index] == 1}
@@ -133,7 +133,7 @@ def create_groups(table, column_number, cares, outputcolumns, multi_output):
 #         (c) information if any reduction was performed in this single
 #             elimination
 
-def reduction_step(groups, n, multi_output):
+def _reduction_step(groups, n, multi_output):
     was_any_reduction = False
     new_groups = []
     for i in range(n + 1):
@@ -178,36 +178,36 @@ def reduction_step(groups, n, multi_output):
[…]
