### 2023-03-24 3ab71c3
Update prime_implicants.py
