+++ paper/paper.bib
@@ -73,6 +73,15 @@
    number = {4},
    pages = {245-264}
 }
+@article{mkrtchyan_et_al_2023,
+   author = {Mkrtchyan, Lusine and Thiem, Alrik and Sebechlebská, Zuzana},
+   year = {2023},
+   title = {Re-establishing a lost connection: Multi-value logic in causal data analysis in social science disciplines},
+   journal = {IEEE Access},
+   volume = {11},
+   pages = {10471-10482},
+   doi = {10.1109/ACCESS.2023.3240094}
+}
 @book{pearl_2009,
    author = {Pearl, J.},
    year = {2009},
@@ -155,10 +164,20 @@
 @article{thiem_et_al_2022,
    author = {Thiem, Alrik and Mkrtchyan, Lusine and Sebechlebská, Zuzana},
    year = {2022},
-   title = {Combinational Regularity Analysis (CORA) — a new method for uncovering complex causation in medical and health research},
+   title = {Combinational Regularity Analysis (CORA) — A new method for uncovering complex causation in medical and health research},
    journal = {BMC Medical Research Methodology},
    volume = {22},
    number = {1},
    pages = {333},
    doi = {10.1186/s12874-022-01800-9}
+}
+@article{thiem_et_al_2023,
+   author = {Thiem, Alrik and Sebechlebská, Zuzana and Mkrtchyan, Lusine},
+   year = {2023},
+   title = {Logic diagrams: A visual tool with untapped potential},
+   journal = {Nature Reviews Methods Primers},
+   volume = {3},
+   number = {1},
+   pages = {21},
+   doi = {10.1038/s43586-023-00210-5}  
 }
