+++ paper/paper.bib
@@ -74,7 +74,7 @@
    pages = {245-264}
 }
 @article{mkrtchyan_et_al_2023,
-   author = {Mkrtchyan, Lusine and Thiem, Alrik and Sebechlebská, Zuzana},
+   author = {Mkrtchyan, L. and Thiem, A. and Sebechlebská, Z.},
    year = {2023},
    title = {Re-establishing a lost connection: Multi-value logic in causal data analysis in social science disciplines},
    journal = {IEEE Access},
@@ -162,7 +162,7 @@
    doi = {10.1371/journal.pone.0233625}
 }
 @article{thiem_et_al_2022,
-   author = {Thiem, Alrik and Mkrtchyan, Lusine and Sebechlebská, Zuzana},
+   author = {Thiem, A. and Mkrtchyan, L. and Sebechlebská, Z.},
    year = {2022},
    title = {Combinational Regularity Analysis (CORA) — A new method for uncovering complex causation in medical and health research},
    journal = {BMC Medical Research Methodology},
@@ -172,7 +172,7 @@
    doi = {10.1186/s12874-022-01800-9}
 }
 @article{thiem_et_al_2023,
-   author = {Thiem, Alrik and Sebechlebská, Zuzana and Mkrtchyan, Lusine},
+   author = {Thiem, A. and Sebechlebská, Z. and Mkrtchyan, L.},
    year = {2023},
    title = {Logic diagrams: A visual tool with untapped potential},
    journal = {Nature Reviews Methods Primers},
