### 2023-03-27 473f246
replace of NaN with - in Colab
