+++ README.md
@@ -27,7 +27,7 @@ It is recommended to install the package into a dedicated virtual environment.
 
 To open CORA with a graphical interface in Google Colab, click the button below:
 
-[![Open in Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/gist/ZuzanaSebb/5bb6e29772dea02c0c562979f52fd27a/cora-1-0-6.ipynb?authuser=1)
+[![Open in Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/gist/ZuzanaSebb/c3ff8e213a974742c4ce85c0f5866be8/cora-1-0-6.ipynb?authuser=1)
 
 ## Usage
 The main features of the package are part of the `OptimizationContext` class, including functions:
