+++ paper/paper.bib
@@ -113,6 +113,16 @@
    pages = {769-771},
    doi = {10.1037/hea0000783}
 }
+@article{thiem_2014,
+   author = {Thiem, A.},
+   year = {2014},
+   title = {Unifying Configurational Comparative Methods: Generalized-Set Qualitative Comparative Analysis},
+   journal = {Sociological Methods & Research},
+   volume = {43},
+   number = {2},
+   pages = {313-337},
+   doi = {10.1177/0049124113500481}
+}
 @book{thiem_2018,
    author = {Thiem, A.},
    year = {2018},
@@ -180,4 +190,14 @@
    number = {1},
    pages = {21},
    doi = {10.1038/s43586-023-00210-5}  
+}
+@article{wang_buck_2012,
+   author = {Wang, B. and Buck, M.},
+   year = {2012},
+   title = {Customizing cell signaling using engineered genetic logic circuits},
+   journal = {Trends in Microbiology},
+   volume = {20},
+   number = {8},
+   pages = {376-384},
+   doi = {10.1016/j.tim.2012.05.001}
 }
