### 2023-03-28 #5019 @Naeemkh
Hi @ZuzanaSebb,
Thank you for the update. The only thing left to complete is the API documentation for Logigram (https://github.com/PoliUniLu/Logigram/issues/1). Please take care of that and let us know once it's done.
