### 2023-04-01 fdc9758
fix in the documentation of Implicant class
