+++ README.md
@@ -27,7 +27,7 @@ It is recommended to install the package into a dedicated virtual environment.
 
 To open CORA with a graphical interface in Google Colab, click the button below:
 
-[![Open in Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/gist/ZuzanaSebb/c3ff8e213a974742c4ce85c0f5866be8/cora-1-0-6.ipynb?authuser=1)
+[![Open in Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/gist/ZuzanaSebb/6c3666591b5e4d969499a65f03f247a9/-cora_1-0-3.ipynb?authuser=1)
 
 ## Usage
 The main features of the package are part of the `OptimizationContext` class, including functions:
