### 2023-04-12 #5019 @ZuzanaSebb
Hi @Naeemkh,
We added documentation to one more exported function in the Logigram package. It's a small package with public API consisting of just two exported functions, all other functions are internal which is now noted with `_` prefix.
