### 2023-04-12 #5019 @ZuzanaSebb
Hello @Naeemkh,
I am sorry that I didn't ping you earlier, but the corresponding [commit](https://github.com/PoliUniLu/Logigram/commit/14fc4bd1a110203da8684f4d8458428bb0d14f38) is 2 weeks old.

Best regards,
Zuzana
