### 2023-04-12 #5019 @Naeemkh
Hi @ZuzanaSebb,
The documentation for an API serves not only its current users but also potential developers, including yourself in the future. Therefore, it is advisable to keep the documentation up-to-date for future reference.
