### 2023-04-14 6cbc545
data_mining fix in input labels
