+++ README.md
@@ -27,8 +27,7 @@ It is recommended to install the package into a dedicated virtual environment.
 
 To open CORA with a graphical interface in Google Colab, click the button below:
 
-[![Open in Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/gist/ZuzanaSebb/4f6f392671b94463e26db41268f80193/-cora_1-0-3.ipynb?authuser=1)
-
+[![Open in Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/gist/ZuzanaSebb/97472c33a9fd74e73c88a2978b2b90c4/-cora_1-0-3.ipynb?authuser=1)
 ## Usage
 The main features of the package are part of the `OptimizationContext` class, including functions:
 - `get_prime_implicants`,
