+++ README.md
@@ -11,7 +11,7 @@ Combinational Regularity Analysis (CORA) is a member of the family of Configurat
 
 In contrast to QCA and CNA, however, CORA has been inspired by switching circuit analysis, a subfield of electrical engineering. INUS structures and switching circuits have much in common because propositional logic - the language of INUS causation - and switching algebra - the language of switching circuit analysis - are operationally equivalent branches of the same underlying Boolean algebra (Lewin and Protheroe 1992). It is therefore no coincidence that one of the first systematic algorithms for Boolean optimization - the Quine-McCluskey algorithm (McCluskey 1956; Quine 1955) - had been co-developed by an analytical philosopher (Willard Van Orman Quine) and an electrical engineer (Edward J. McCluskey).
 
-Most importantly, CORA is currently the only CCM able to analyze INUS structures that simultaneously feature simple as well as complex effects (e.g., *y* **and** **not** *z*, **not** *y* **and** *z*, *y* **and** *z*). CORA can process such structures even in multi-value form (Mkrtchyan *et al*. 2023). In addition, CORA offers a configurational version of Occam's Razor: a data-mining approach to solution building that reduces model ambiguities by keeping the number of required variables for finding a solution at a minimum. Lastly, CORA includes a lean yet powerful visualization module called LOGIGRAM, with which two-level logic diagrams can be produced from any (system of) Boolean or multi-value function(s) in disjunctive normal form. Logic diagrams considerably outperfrom Venn diagrams, which are standardly used in QCA, when it comes to the visualization and interpretability of INUS structures (Thiem *et al*. 2023).
+Most importantly, CORA is currently the only CCM able to analyze INUS structures that simultaneously feature simple as well as complex effects (e.g., *y* **and** **not** *z*, **not** *y* **and** *z*, *y* **and** *z*). CORA can process such structures even in multi-value form (Mkrtchyan *et al*. 2023). In addition, CORA offers a configurational version of Occam's Razor: a data-mining approach to solution building that reduces model ambiguities by keeping the number of required variables for finding a solution at a minimum. Lastly, CORA includes a lean yet powerful visualization module called LOGIGRAM, with which two-level logic diagrams can be produced from any (system of) Boolean or multi-value function(s) in disjunctive normal form. Logic diagrams considerably outperfrom Venn diagrams, which are often used in QCA, when it comes to the representation and interpretability of INUS structures (Thiem *et al*. 2023).
 
 ## Installation
 
@@ -49,10 +49,10 @@ df = pd.DataFrame([[1,1,0,1],
                    [0,1,0,1]], columns=["A","B","C","OUT"])
 
 context = OptimizationContext(data = df, output_labels = ["OUT"])
-PIs = context.get_prime_implicants() #result: (B, c, #a)
+PIs = context.get_prime_implicants() # result: (B, c, #a); essential prime implicants are marked by prefixed hashtags
 irredundant_solutions = context.get_irredundant_sums() # result: [M1: #a + B, M2: #a + c]
 ```
-The data-mining approach that searches for feasible tuples of solution-generating inputs is another feature of the package.
[…]
