+++ pyproject.toml
@@ -8,8 +8,8 @@ readme = "README.md"
 [tool.poetry.dependencies]
 python = ">=3.7.1,<3.11"
 nativepetric = {git = "https://github.com/PoliUniLu/petric.git"}
-numpy = "==1.21.2"
-pandas = "==1.3.4"
+numpy = "==1.22"
+pandas = "==1.3.5"
 
 [tool.poetry.dev-dependencies]
 parameterized = "^0.8.1"
