+++ pyproject.toml
@@ -9,7 +9,7 @@ readme = "README.md"
 python = ">=3.8,<3.11"
 nativepetric = {git = "https://github.com/PoliUniLu/petric.git"}
 numpy = "==1.22"
-pandas = "==1.3.5"
+pandas = "==1.5.3"
 
 [tool.poetry.dev-dependencies]
 parameterized = "^0.8.1"
