### 2023-04-21 8436762
fix in prime_implicants.py
