### 2023-04-21 #5019 @ZuzanaSebb
Hi @vissarion,
Here are the release versions of Cora: https://github.com/PoliUniLu/cora/releases/tag/v1.0.0 and Logigram: https://github.com/PoliUniLu/Logigram/releases
The zenodo archive links: https://zenodo.org/record/7853729#.ZELywexBxJU
https://zenodo.org/record/7853733#.ZELzEuxBxJU
