+++ README.md
@@ -78,7 +78,7 @@ result # print(result.to_markdown())
 |  5 | ['C', 'D']    |               1 |        0.75 |           1 |    0.75 |
 
 ```
-To access more examples see the `/examples` folder or follow [![Open in Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/gist/ZuzanaSebb/800d5d1fce3da66127369ccc7511a122/-cora_1-0-3.ipynb?authuser=1)
+To access more examples see the `/examples` folder or follow [![Open in Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/gist/ZuzanaSebb/5bd80c5cc7ab027ade9d589e3fb04b2d/-cora_1-0-3.ipynb?authuser=1)
 
 
 
