### 2023-04-23 86c7b91
fix in data mining
