### 2023-04-24 #5019 @vissarion
@ZuzanaSebb since we cannot assign two versions and archives to the paper, I propose to use the version and archive of CORA package (which is the main package) and then you can refer to LOGIGRAM in the description of zenodo package by also adding a link to the LOGIGRAM's zenodo page.

Alternatively you may create a zenodo archive with both CORA and LOGIGRAM. Please note that the title of the zenodo archive (including authors names, affiliation and order) should be **exactly the same** as the one of the paper.
