### 2023-04-24 #5019 @vissarion
Actually an advantage of the second option mentioned above is that we reduce confusion to the reader since the title "CORA and LOGIGRAM: A Duo of Python Packages for Combinational Regularity Analysis (CORA)" refer to both packages that are expected to be found in archive.
