+++ README.md
@@ -28,7 +28,7 @@ It is recommended to install the package into a dedicated virtual environment.
 ## Google Colab
 
 To open CORA with a graphical interface in Google Colab, click the button below:
-[![Open in Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/gist/ZuzanaSebb/aae880e37c5d0f32f9908daaba5bc6a0/-cora_1-0-3.ipynb?authuser=1)
+[![Open in Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/gist/ZuzanaSebb/7b53b37c55c4e26c5582d13c0558691f/-cora_1-0-3.ipynb)
 
 ## Usage
 The main features of the package are part of the `OptimizationContext` class, including functions:
