+++ cora/prime_implicants.py
@@ -1094,7 +1094,7 @@ class OptimizationContext:
         if not self.multi_output:
             irr_sums = self.get_irredundant_sums()
             new_cols = ["M" + str(x.index) for x in irr_sums]
-            cov_per_impl = [x.impl_cov_score1() for x in irr_sums]
+            cov_per_impl = [x.impl_cov_score() for x in irr_sums]
 
         else:
             irr_systems = self.get_irredundant_systems()
@@ -1110,33 +1110,6 @@ class OptimizationContext:
 
         self.details = df_implicant
         return self.details
-    def pi_details_experimental(self):
-
-
-        prime_implicants = self.get_prime_implicants()
-
-        cov_x = [(x.implicant,
-                  round(x.coverage_score(), 2),
-                  round(x.inclusion_score(), 2)) for x in prime_implicants]
-
-        if not self.multi_output:
-            irr_sums = self.get_irredundant_sums()
-            new_cols = ["M" + str(x.index) for x in irr_sums]
-            cov_per_impl = [x.impl_cov_score() for x in irr_sums]
-
-        else:
-            irr_systems = self.get_irredundant_systems()
-            new_cols = ["S" + str(x.index) for x in irr_systems]
-            cov_per_impl = [x.impl_cov_score1() for x in irr_systems]
-
-        df_implicant = pd.DataFrame(cov_x, columns=['PI', 'Cov.r', 'Inc.'])
-        for f_cov_per_impl, f_label in zip(cov_per_impl, new_cols):
-            tmp = [round(f_cov_per_impl[x.implicant], 2) if x.implicant
-                                            in f_cov_per_impl.keys() else None
-                   for x in prime_implicants]
-            df_implicant[f_label] = tmp
-
-        return  df_implicant
 
     def get_irredundant_systems(self):
         """
