### 2023-04-26 75217c3
fix in coverages with respect to the corresponding implicants
