### 2023-04-26 #5019 @ZuzanaSebb
Hi @vissarion ,
I created a new zenodo archive containing both repositories,
https://zenodo.org/record/7869083#.ZEmD3exBxJU
where I also updated the version of Cora to v1.0.1.
https://github.com/PoliUniLu/cora/releases/tag/v1.0.1
