### 2023-04-27 PoliUniLu/cora#18 @vissarion
[JOSS Review] References
There are some minor issues in references of the joss paper.
- cna, QCA and QCAPro CRAN packages could have a clickable link to their cran page
- Ragin, C. C., & Davey, S. (2019). fs/QCA: fuzzy-set/Qualitative Comparative Analysis, the link to the webpage doesn't seem to work
