+++ paper/paper.bib
@@ -39,7 +39,8 @@
    year = {2019},
    title = {Tosmana: Tool for small-n analysis, version 1.61 [computer program]},
    publisher = {University of Trier},
-   address = {Trier}
+   address = {Trier},
+   doi = {10.978.395710/3505}
 }
 @article{delorenzo_schmidt_2018,
    author = {de Lorenzo, V. and Schmidt, M.},
@@ -127,7 +128,8 @@
    author = {Thiem, A.},
    year = {2018},
    title = {{QCApro}: Advanced functionality for performing and evaluating {Qualitative Comparative Analysis}, {R} package version 1.1-2},
-   publisher = {http://cran.r-project.org/package=QCApro}
+   publisher = {http://cran.r-project.org/package=QCApro},
+   doi = {}
 }
 @incollection{thiem_2022a,
    author = {Thiem, A.},
