+++ paper/paper.bib
@@ -128,8 +128,7 @@
    author = {Thiem, A.},
    year = {2018},
    title = {{QCApro}: Advanced functionality for performing and evaluating {Qualitative Comparative Analysis}, {R} package version 1.1-2},
-   publisher = {http://cran.r-project.org/package=QCApro},
-   doi = {}
+   publisher = {http://cran.r-project.org/package=QCApro}
 }
 @incollection{thiem_2022a,
    author = {Thiem, A.},
