### 2023-04-28 #5019 @vissarion
If you mean QCApro, it is a CRAN package so as far as I know there is no assigned DOI.
Please also have a look at https://github.com/PoliUniLu/cora/issues/18
