+++ paper/paper.bib
@@ -2,7 +2,7 @@
    author = {Ambuehl, M. and Baumgartner, M.},
    year = {2022},
    title = {{cna}: Causal modeling with {Coincidence Analysis}, version 3.4.0},
-   publisher = {http://cran.r-project.org/package=cna}
+   doi = {http://cran.r-project.org/package=cna}
 }
 @article{baumgartner_2009,
    author = {Baumgartner, M.},
@@ -55,7 +55,7 @@
    author = {Dusa, A.},
    year = {2022},
    title = {{QCA}: {Qualitative Comparative Analysis}, {R} package version 3.17},
-   publisher = {https://cran.r-project.org/package=QCA}
+   doi = {https://cran.r-project.org/package=QCA}
 }
 @book{lewin_protheroe_1992,
    author = {Lewin, D. and Protheroe, D.},
@@ -102,7 +102,7 @@
    author = {Ragin, C.C. and Davey, S.},
    year = {2019},
    title = {{fs/QCA}: {fuzzy-set/Qualitative Comparative Analysis}, version 3.1b [Computer Program]},
-   publisher = {http://www.socsci.uci.edu/~cragin/fsQCA/software.shtml}
+   doi = {http://www.socsci.uci.edu/~cragin/fsQCA/software.shtml}
 }
 @article{suls_green_2019,
    author = {Suls, J. and Green, P.A.},
@@ -118,7 +118,7 @@
    author = {Thiem, A.},
    year = {2014},
    title = {Unifying Configurational Comparative Methods: Generalized-Set Qualitative Comparative Analysis},
-   journal = {Sociological Methods & Research},
+   journal = {Sociological Methods \& Research},
    volume = {43},
    number = {2},
    pages = {313-337},
@@ -128,7 +128,7 @@
    author = {Thiem, A.},
    year = {2018},
    title = {{QCApro}: Advanced functionality for performing and evaluating {Qualitative Comparative Analysis}, {R} package version 1.1-2},
-   publisher = {http://cran.r-project.org/package=QCApro}
+   doi = {http://cran.r-project.org/package=QCApro}
 }
 @incollection{thiem_2022a,
    author = {Thiem, A.},
