+++ paper/paper.bib
@@ -1,8 +1,8 @@
-@book{ambuhl_baumgartner_2022,
+@misc{ambuhl_baumgartner_2022,
    author = {Ambuehl, M. and Baumgartner, M.},
    year = {2022},
    title = {{cna}: Causal modeling with {Coincidence Analysis}, version 3.4.0},
-   doi = {http://cran.r-project.org/package=cna}
+   url = {http://cran.r-project.org/package=cna}
 }
 @article{baumgartner_2009,
    author = {Baumgartner, M.},
@@ -34,13 +34,13 @@
    pages = {279-311},
    doi = {10.1177/0049124117701487}
 }
-@book{cronqvist_2019,
+@misc{cronqvist_2019,
    author = {Cronqvist, L.},
    year = {2019},
    title = {Tosmana: Tool for small-n analysis, version 1.61 [computer program]},
    publisher = {University of Trier},
    address = {Trier},
-   doi = {10.978.395710/3505}
+   url = {https://www.tosmana.net}
 }
 @article{delorenzo_schmidt_2018,
    author = {de Lorenzo, V. and Schmidt, M.},
@@ -51,11 +51,11 @@
    pages = {170-180},
    doi = {10.1016/j.nbt.2017.05.001}
 }
-@book{dusa_2022,
+@misc{dusa_2022,
    author = {Dusa, A.},
    year = {2022},
    title = {{QCA}: {Qualitative Comparative Analysis}, {R} package version 3.17},
-   doi = {https://cran.r-project.org/package=QCA}
+   url = {https://cran.r-project.org/package=QCA}
 }
 @book{lewin_protheroe_1992,
    author = {Lewin, D. and Protheroe, D.},
@@ -98,11 +98,11 @@
    publisher = {University of California Press},
    address = {Berkeley}
 }
-@book{ragin_davey_2019,
+@misc{ragin_davey_2019,
    author = {Ragin, C.C. and Davey, S.},
    year = {2019},
    title = {{fs/QCA}: {fuzzy-set/Qualitative Comparative Analysis}, version 3.1b [Computer Program]},
-   doi = {http://www.socsci.uci.edu/~cragin/fsQCA/software.shtml}
+   url = {http://www.socsci.uci.edu/~cragin/fsQCA/software.shtml}
 }
 @article{suls_green_2019,
    author = {Suls, J. and Green, P.A.},
@@ -124,11 +124,11 @@
    pages = {313-337},
    doi = {10.1177/0049124113500481}
 }
-@book{thiem_2018,
+@misc{thiem_2018,
    author = {Thiem, A.},
    year = {2018},
    title = {{QCApro}: Advanced functionality for performing and evaluating {Qualitative Comparative Analysis}, {R} package version 1.1-2},
-   doi = {http://cran.r-project.org/package=QCApro}
+   url = {http://cran.r-project.org/package=QCApro}
 }
 @incollection{thiem_2022a,
    author = {Thiem, A.},
