+++ paper/paper.bib
@@ -175,7 +175,7 @@
 @article{thiem_et_al_2022,
    author = {Thiem, A. and Mkrtchyan, L. and Sebechlebská, Z.},
    year = {2022},
-   title = {Combinational Regularity Analysis (CORA) — A new method for uncovering complex causation in medical and health research},
+   title = {Combinational Regularity Analysis (CORA) — a new method for uncovering complex causation in medical and health research},
    journal = {BMC Medical Research Methodology},
    volume = {22},
    number = {1},
@@ -192,6 +192,16 @@
    pages = {21},
    doi = {10.1038/s43586-023-00210-5}  
 }
+@article{tison_1967,
+   author = {Tison, Pierre},
+   year = {1967},
+   title = {Generalization of consensus theory and application to the minimization of Boolean functions},
+   journal = {IEEE Transactions on Electronic Computers},
+   volume = {16},
+   number = {4},
+   pages = {446-456},
+   doi = {10.1109/PGEC.1967.264648}  
+}
 @article{wang_buck_2012,
    author = {Wang, B. and Buck, M.},
    year = {2012},
@@ -201,4 +211,14 @@
    number = {8},
    pages = {376-384},
    doi = {10.1016/j.tim.2012.05.001}
+}
+@article{whitaker_et_al_2020,
+   author = {Whitaker, Rebecca Garr and Sperber, Nina and Baumgartner, Michael and Thiem, Alrik and Cragun, Deborah and Damschroder, Laura and Miech, Edward J. and Slade, Alecia and Birken, Sarah},
+   year = {2020},
+   title = {Coincidence analysis: A new method for causal inference in implementation science},
+   journal = {Implementation Science},
+   volume = {15},
+   number = {1},
+   pages = {108},
+   doi = {10.1186/s13012-020-01070-3}
 }
