### 2023-05-07 #5019 @ZuzanaSebb
Hi @vissarion,
We made the tag release cora [v1.0.2](https://github.com/PoliUniLu/cora/releases/tag/v1.0.2)
and the zenodo archive version: https://zenodo.org/record/7898680#.ZFgJvexBxJU
